% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access.R
\name{add_access_rules}
\alias{add_access_rules}
\title{access.R}
\usage{
add_access_rules(sysmeta)
}
\arguments{
\item{sysmeta}{(SystemMetadata) The SystemMetadata to add rules to.}
}
\value{
The modified SystemMetadata object
}
\description{
Utility functions for getting and setting access rules for DataONE objects.
Add access rules to the sysmeta object
}
\details{
This is a function because I add a set of standard set of access rules to
every object and the access rules don't differ across objects.
}

