% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{add_additional_identifiers}
\alias{add_additional_identifiers}
\title{Add a set of additional identifiers to an EML document.}
\usage{
add_additional_identifiers(path, identifiers)
}
\arguments{
\item{path}{(character) Path to the EML document.}

\item{identifiers}{(character) Set of identifiers to add.}
}
\value{
(character) Path to the modified document.
}
\description{
Add a set of additional identifiers to an EML document.
}
