% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sysmeta.R
\name{add_admin_group_access}
\alias{add_admin_group_access}
\title{Adds access to the given System Metadata for the arctic-data-admins group}
\usage{
add_admin_group_access(sysmeta)
}
\arguments{
\item{sysmeta}{}
}
\description{
Adds access to the given System Metadata for the arctic-data-admins group
}

