% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{add_dummy_prov}
\alias{add_dummy_prov}
\title{Add prov to a dummy package}
\usage{
add_dummy_prov(mn, rm_pid)
}
\arguments{
\item{mn}{member node (the ADC test node)}

\item{rm_pid}{resource map identifier}
}
\description{
Adds provenance information to a dummy package for testing
}
