% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml.R
\name{add_methods_step}
\alias{add_methods_step}
\title{Adds a step to the methods document}
\usage{
add_methods_step(doc, title, description)
}
\arguments{
\item{doc}{(eml) The EML document to add the method step to.}

\item{title}{(character) The title of the method step.}

\item{description}{(character) The description of the method.}
}
\value{
(eml) The modified EML document
}
\description{
Adds a step to the methods document
}
\examples{
\dontrun{
# Add a method step to a fresh document
my_doc <- new("eml")
my_doc <- add_methods_step(my_doc, "some method", "how I did the method")
}
}

