% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml.R
\name{add_other_entities}
\alias{add_other_entities}
\title{Creates and adds EML otherEntity elements to an existing EML document.}
\usage{
add_other_entities(mn, path, pids)
}
\arguments{
\item{mn}{(MNode) The Member Node the objects exist on.}

\item{path}{(character) The location on disk of the EML file.}

\item{pids}{(character) One or more PIDs for the objects.}
}
\value{
(character) The path to the updated EML file.
}
\description{
Creates and adds EML otherEntity elements to an existing EML document.
}

