% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{add_string_to_title}
\alias{add_string_to_title}
\title{Adds a string to the title element in the given file.}
\usage{
add_string_to_title(path, string)
}
\arguments{
\item{path}{(character) Path to the XML file to edit.}

\item{string}{(character) The new value.}
}
\description{
Adds a string to the title element in the given file.
}

