% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{change_eml_name}
\alias{change_eml_name}
\title{Utility function to extract a name string from an XML individualName node,
parse it into tokens,and reformat the individualName with new children nodes}
\usage{
change_eml_name(party)
}
\arguments{
\item{party}{the XML node containing a subclass of eml:ResponsibleParty}
}
\value{
the modified XML node
}
\description{
Utility function to extract a name string from an XML individualName node,
parse it into tokens,and reformat the individualName with new children nodes
}
