% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{check_format}
\alias{check_format}
\title{Check that the given format is valid}
\usage{
check_format(format)
}
\arguments{
\item{format}{(character) The format ID to check.}
}
\value{
(logical) Whether or not the format was valid.
}
\description{
Validity is determined by the given format being found in the list on
\url{https://cn.dataone.org/cn/v2/formats}.
}
