% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml.R
\name{clear_methods}
\alias{clear_methods}
\title{Clear all methods from the document.}
\usage{
clear_methods(doc)
}
\arguments{
\item{doc}{(eml) The document to clear methods from.}
}
\value{
(eml) The modified document.
}
\description{
Clear all methods from the document.
}
\examples{
# First we create a new EML document and add a method
doc <- new("eml")
doc <- add_method_step(doc, "some lab method", "this is how we did it")
doc

# Then we clear it
clear_methods(doc)
}

