% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sysmeta.R
\name{clear_replication_policy}
\alias{clear_replication_policy}
\title{Clear the replication policy from a System Metadata object}
\usage{
clear_replication_policy(sysmeta)
}
\arguments{
\item{sysmeta}{(SystemMetadata) The System Metadata object to clear the replication policy of.}
}
\value{
(SystemMetadata) The modified System Metadata object.
}
\description{
Clear the replication policy from a System Metadata object
}
