% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{create_dummy_attributes_dataframe}
\alias{create_dummy_attributes_dataframe}
\title{Create dummy attributes data frame}
\usage{
create_dummy_attributes_dataframe(numberAttributes, factors = NULL)
}
\arguments{
\item{numberAttributes}{(integer) Number of attributes to be created in the table}

\item{factors}{(character) Optional vector of factor names to include.}
}
\value{
(data.frame) Data frame of attributes
}
\description{
Create dummy attributes data frame
}
\examples{
\dontrun{
# Create dummy attribute dataframe with 6 attributes and 1 factor
attributes <- create_dummy_attributes_dataframe(6, c("Factor1", "Factor2"))
}
}
