% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{create_dummy_enumeratedDomain_dataframe}
\alias{create_dummy_enumeratedDomain_dataframe}
\title{Create dummy enumeratedDomain data frame}
\usage{
create_dummy_enumeratedDomain_dataframe(factors)
}
\arguments{
\item{factors}{(character) Optional vector of factor names to include.}
}
\description{
Create dummy enumeratedDomain data frame
}
