% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{create_dummy_metadata}
\alias{create_dummy_metadata}
\title{Create a test metadata object}
\usage{
create_dummy_metadata(mn, data_pids = NULL)
}
\arguments{
\item{mn}{(MNode) The Member Node.}

\item{data_pids}{(character) Optional. PIDs for data objects the metadata documents.}
}
\value{
(character) The PID of the published metadata document.
}
\description{
Create a test EML metadata object.
}
\examples{
\dontrun{
# Set environment
cn <- CNode("STAGING2")
mn <- getMNode(cn,"urn:node:mnTestKNB")
pid <- create_dummy_metadata(mn)
}
}
