% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{create_dummy_package_full}
\alias{create_dummy_package_full}
\title{Create dummy package with fuller metadata}
\usage{
create_dummy_package_full(mn, title = "A Dummy Package")
}
\arguments{
\item{mn}{(MNode) The Member Node.}

\item{title}{(character) Optional. Title of package. Defaults to "A Dummy Package".}
}
\value{
(list) The PIDs for all elements in the data package.
}
\description{
Creates a fuller package than \code{\link[=create_dummy_package]{create_dummy_package()}}
but is otherwise based on the same concept. This dummy
package includes multiple data objects, responsible parties,
geographic locations, method steps, etc.
}
