% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{create_dummy_parent_package}
\alias{create_dummy_parent_package}
\title{Create a test parent package.}
\usage{
create_dummy_parent_package(mn, children)
}
\arguments{
\item{mn}{(MNode) The Member Node.}

\item{children}{(character) Child package (resource maps) PIDs.}
}
\value{
pid (character) Named character vector of PIDs including parent package and child package pids.
}
\description{
Create a test parent package.
}
\examples{
\dontrun{
# Set environment
}
}
