% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inserting.R
\name{create_from_folder}
\alias{create_from_folder}
\title{inserting.R}
\usage{
create_from_folder(mn, path, data_pids = NULL)
}
\arguments{
\item{mn}{(MNode) The Member Node to create the packages on.}

\item{path}{(character) The path to the folder containing the files.}

\item{data_pids}{(character) Optional. Manually specify the PIDs of data. This is useful if data were inserted outside this function and you want to re-use those objects.}
}
\value{
(list) All of the PIDs created.
}
\description{
A set of utilities for inserting packages from files and folders on disk.
Create a package from a folder containing an ISO package (legacy)
}
\details{
This function  handles the process of inserting the original ISO package
and updating it with an EML package.

Note: This only works for Gateway packages right now.
}

