% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packaging.R
\name{create_object}
\alias{create_object}
\title{Create an object from a row of the inventory.}
\usage{
create_object(file, sysmeta, base_path, mn)
}
\arguments{
\item{file}{(data.frame)A row from the inventory.}

\item{sysmeta}{(SystemMetadata) The file's sysmeta.}

\item{base_path}{(character)}

\item{mn}{(MNode)}
}
\description{
Create an object from a row of the inventory.
}
