% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editing.R
\name{create_resource_map}
\alias{create_resource_map}
\title{Create a resource map.}
\usage{
create_resource_map(mn, metadata_pid, data_pids = NULL, child_pids = NULL)
}
\arguments{
\item{child_pids}{}
}
\value{
The created resource map's PID (character)
}
\description{
Similar to but different than update_resource_map in that it uses a create
call instead of an update call.
}

