% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packaging.R
\name{create_sysmeta}
\alias{create_sysmeta}
\title{Create a sysmeta object.}
\usage{
create_sysmeta(file, base_path, submitter, rights_holder)
}
\arguments{
\item{file}{(data.frame) A single row from the inventory.}

\item{base_path}{(character) The path prefix to use with the contents of `file[1,"filename]` that
will be used to locate the file on disk.}

\item{submitter}{(character) The submitter DN string for the object.}

\item{rights_holder}{(character) The rights holder DN string for the object.}
}
\value{
The sysmeta object (dataone::SystemMetadata)
}
\description{
This is a wrapper function around the constructor for a
dataone::SystemMetadata object.
}

