% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packaging.R
\name{determine_child_pids}
\alias{determine_child_pids}
\title{Calculate a set of child PIDs for a given package.}
\usage{
determine_child_pids(inventory, package)
}
\arguments{
\item{package}{}
}
\description{
Calculate a set of child PIDs for a given package.
}
