% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology.R
\name{eml_adcad_annotation}
\alias{eml_adcad_annotation}
\title{Given a term from the ADC Academic Disciplines (ADCAD) ontology, produce the corresponding annotation}
\usage{
eml_adcad_annotation(valueLabel)
}
\arguments{
\item{valueLabel}{(character) One of the disciplines found in
\href{https://bioportal.bioontology.org/ontologies/OBOE/?p=classes&conceptid=root}{ADCAD}}
}
\value{
list - a formatted EML annotation
}
\description{
Reduces the amount of copy pasting needed
}
\examples{
eml_ecso_annotation("latitude coordinate")
}
