% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml.R
\name{eml_add_csv_headers}
\alias{eml_add_csv_headers}
\title{Add Number of Header Lines to Physical}
\usage{
eml_add_csv_headers(mn, lines, pid)
}
\arguments{
\item{mn}{(MNode) The Member Node to publish the object to.}

\item{lines}{(character) number of rows to be ignored before the row containing column names}

\item{pid}{pid of csv file
de}
}
\value{
Returns physical with number of header lines field
}
\description{
When a csv file contains multiple header lines, this function can be used to account for that in the dataTable physical
}
\examples{
\dontrun{
# Set environment

# Generate EML physical section for an object in a data package with header lines'

library(EML)
library(dataone)

cn_staging <- CNode('STAGING')
mn_test <- getMNode(cn_staging,'urn:node:mnTestARCTIC')

pkg <- arcticdatautils::create_dummy_package_full(mn_test, "A Dummy Package")

doc <- EML::read_eml(getObject(mn_test, pkg$metadata))

pid <- pkg$data[1]

lines <- sample(1:100, 1)

phys <- arcticdatautils::eml_add_csv_headers(mn_test, lines, pid)

doc$dataset$dataTable[[1]]$physical <- phys

eml_validate(doc)

}
}
\author{
Rachel Sun rachelsun@ucsb.edu
}
