% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml.R
\name{eml_add_distribution}
\alias{eml_add_distribution}
\title{Add distribution information to EML}
\usage{
eml_add_distribution(doc, identifier)
}
\arguments{
\item{doc}{(emld) An EML document}

\item{identifier}{(character) A pre-issued, unassigned identifier (as from \code{dataone::generateIdentifier()})}
}
\value{
doc (emld) An EML document with distribution added
}
\description{
Adds a landing page URL to the dataset, and corrects the metadata identifier
by replacing the existing identifier with that which is passed. Note that this
function constructs landing page URLs for the Arctic Data Center only and will not work
correctly on other repositories.
}
\examples{
\dontrun{
library(EML)
d1c <- dataone::D1Client("STAGING", "mnTestARCTIC")
# read in any EML document
doc <- read_eml(system.file("extdata/strix-pacific-northwest.xml", package="dataone"))
# generate a doi
id <- generateIdentifier(d1c@mn, "doi")
doc <- eml_add_distribution(doc, id)
}

}
