% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml.R
\name{eml_add_entity_system}
\alias{eml_add_entity_system}
\title{Add system information to entities}
\usage{
eml_add_entity_system(doc)
}
\arguments{
\item{doc}{(emld) An EML document}
}
\value{
(emld) An EML document
}
\description{
This function adds system information to entities in a document
}
\examples{
\dontrun{
# Add publisher information to an existing document
doc <- eml_add_entity_system(doc)
}
}
