% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml.R
\name{eml_add_other_entities}
\alias{eml_add_other_entities}
\title{Add new otherEntity elements to an EML document from a table}
\usage{
eml_add_other_entities(doc, entities,
  resolve_base = "https://cn.dataone.org/cn/v2/resolve/")
}
\arguments{
\item{doc}{(eml) An EML document}

\item{entities}{(data.frame) A data.frame with columns path, pid, and
format_id}

\item{resolve_base}{(character) Optional. Specify a DataONE CN resolve base
URI which will be used for serializing download URLs into the EML. Most users
 should not override the default value.}
}
\value{
(eml) The modified EML document.
}
\description{
Add new otherEntity elements to an EML document from a table
}
\examples{
# Create entities from files on disk
paths <- list.files(system.file("", package = "arcticdatautils"), full.names = TRUE) # Get full paths to some files
pids <- vapply(paths, function(x) { paste0("urn:uuid:", uuid::UUIDgenerate()) }, "") # Generate some UUID PIDs
format_ids <- guess_format_id(paths) # Try to guess format IDs, you should check this afterwards
entity_df <- data.frame(path = paths, pid = pids, format_id = format_ids, stringsAsFactors = FALSE)

doc <- new("eml")
doc <- eml_add_other_entities(doc, entity_df)

# Read in a CSV containing the info
\dontrun{
  entity_df <- read.csv("./my_entities.csv", stringsAsFactors = FALSE)
  doc <- new("eml")
  doc <- eml_add_other_entities(doc, entity_df)
}
}
