% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml.R
\name{eml_add_publisher}
\alias{eml_add_publisher}
\title{Add publisher information to EML document}
\usage{
eml_add_publisher(doc)
}
\arguments{
\item{doc}{(emld) An EML document}
}
\value{
(emld) An EML document
}
\description{
This function adds Arctic Data Center publisher information to an EML document
}
\examples{
\dontrun{
# Add publisher information to an existing document
doc <- eml_add_publisher(doc)
}
}
