% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml.R
\name{eml_address}
\alias{eml_address}
\title{Create an EML address element.}
\usage{
eml_address(delivery_points, city, administrative_area, postal_code)
}
\arguments{
\item{delivery_points}{(character) One or more delivery points.}

\item{city}{(character) City}

\item{administrative_area}{(character) Administrative area}

\item{postal_code}{(character) Postal code}
}
\value{
(address) An EML address object.
}
\description{
Create an EML address element.
}

