% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml.R
\name{eml_arcrc_add_annotation}
\alias{eml_arcrc_add_annotation}
\title{Add an Arctic Report Card annotation to a dataset}
\usage{
eml_arcrc_add_annotation(doc, property, label)
}
\arguments{
\item{doc}{(emld) An EML document}

\item{property}{(character) One of two properties: "isAbout" for key variables or "influenced" for essay topics}

\item{label}{(character) One or more labels in title case from the ADCAD ontology.}
}
\value{
doc (emld) An EML document with annotation added
}
\description{
Creates an annotation from the Arctic Report Card ontology
\href{https://bioportal.bioontology.org/ontologies/ARCRC/?p=summary}{here}
and inserts the annotation into the EML document \code{doc} while retaining any existing
annotations such as the sensitivity annotations or dataset categorization. For a
list of available essay topics or key variables, see link above.
}
\examples{
library(EML)
# read in any EML document
doc <- read_eml(system.file("extdata/strix-pacific-northwest.xml", package="dataone"))
# add the dataset categories
doc <- eml_arcrc_add_annotation(doc, "isAbout", c("sea ice thickness", "sea surface temperature"))

}
