% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology.R
\name{eml_arcrc_key_variable_annotation}
\alias{eml_arcrc_key_variable_annotation}
\title{Given a key variable from the Arctic Report Card (ARCRC) ontology, produce the corresponding annotation}
\usage{
eml_arcrc_key_variable_annotation(valueLabel)
}
\arguments{
\item{valueLabel}{(character) One of the key variables found in
\href{https://bioportal.bioontology.org/ontologies/ARCRC/?p=classes&conceptid=http\%3A\%2F\%2Fpurl.dataone.org\%2Fodo\%2FARCRC_00000040}{ARCRC}}
}
\value{
list - a formatted EML annotation
}
\description{
Reduces the amount of copy pasting needed
}
\examples{
eml_arcrc_key_variable_annotation("age of sea ice")
}
