% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml.R
\name{eml_categorize_dataset}
\alias{eml_categorize_dataset}
\title{Categorize a dataset with an annotation}
\usage{
eml_categorize_dataset(doc, discipline)
}
\arguments{
\item{doc}{(emld) An EML document}

\item{discipline}{(character) One or more disciplines in title case from the ADCAD ontology.}
}
\value{
doc (emld) An EML document with annotation added
}
\description{
Creates an annotation from the ADC Academic Disciplines ontology
\href{https://bioportal.bioontology.org/ontologies/ADCAD/?p=classes&conceptid=root}{here}
and inserts the annotation into the EML document \code{doc} while retaining any existing
annotations such as the sensitivity annotations. For a list of available disciplines,
see link above.
}
\examples{
library(EML)
# read in any EML document
doc <- read_eml(system.file("extdata/strix-pacific-northwest.xml", package="dataone"))
# add the dataset categories
doc <- eml_categorize_dataset(doc, c("Soil Science", "Ecology"))

}
