% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml.R
\name{eml_creator}
\alias{eml_creator}
\title{Create an EML creator}
\usage{
eml_creator(...)
}
\arguments{
\item{...}{Arguments passed on to eml_party}
}
\value{
(creator) The new creator
}
\description{
See \code{\link{eml_party}} for details.
}
\examples{
eml_creator("test", "user", email = "test@user.com")
}
