% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology.R
\name{eml_ecso_annotation}
\alias{eml_ecso_annotation}
\title{Given a an annotation from the ECSO ontology, produce the corresponding annotation}
\usage{
eml_ecso_annotation(valueLabel)
}
\arguments{
\item{valueLabel}{(character) the label for the annotation found in
\href{https://bioportal.bioontology.org/ontologies/ECSO/?p=classes&conceptid=root}{ECSO}}
}
\value{
list - a formatted EML annotation
}
\description{
Reduces the amount of copy pasting needed
}
\examples{
eml_ecso_annotation("latitude coordinate")
}
