% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml.R
\name{eml_get_raster_metadata}
\alias{eml_get_raster_metadata}
\title{Get raster info from a file on disk}
\usage{
eml_get_raster_metadata(path, coord_name = NULL, attributes)
}
\arguments{
\item{path}{(char) Path to a raster file}

\item{coord_name}{(char) horizCoordSysDef name}

\item{attributes}{(attributeList) attribute list for raster}
}
\description{
This function populates a spatialRaster element with the
required elements by reading a local raster file in. The
\code{coord_name} argument can be found by examining the data.frame
that \code{get_coord_list()} returns against the proj4string of the
raster file.
}
