% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml.R
\name{eml_individual_name}
\alias{eml_individual_name}
\title{Create an EML individualName section}
\usage{
eml_individual_name(first, last)
}
\arguments{
\item{first}{(character) First name.}

\item{last}{(character) Last name.}
}
\value{
(individualName) The new individualName section
}
\description{
Create an EML individualName section
}

