% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml.R
\name{eml_metadata_provider}
\alias{eml_metadata_provider}
\title{Create an EML metadataProvider}
\usage{
eml_metadata_provider(given_names, sur_name, organization = NULL,
  email = NULL, phone = NULL, address = NULL)
}
\arguments{
\item{given_names}{(character) One or more given (first) names.}

\item{sur_name}{(character) A sur (last) name.}

\item{organization}{(character) One or more organization names.}

\item{email}{(character) An email address.}

\item{phone}{(character) A phone number.}

\item{address}{(address) An object of type 'address' (EML).}
}
\value{
(metadataProvider) The new metadataProvider.
}
\description{
Create an EML metadataProvider
}
\examples{
eml_metadata_provider("test", "user", "test@user.com")
}

