% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{eml_party}
\alias{eml_party}
\title{Create an EML party}
\usage{
eml_party(
  type = "associatedParty",
  given_names = NULL,
  sur_name = NULL,
  organization = NULL,
  position = NULL,
  email = NULL,
  phone = NULL,
  address = NULL,
  userId = NULL,
  role = NULL
)
}
\arguments{
\item{type}{(character) The type of party (e.g. 'contact').}

\item{given_names}{(character) The party's given name(s).}

\item{sur_name}{(character) The party's surname.}

\item{organization}{(character) The party's organization name.}

\item{position}{(character) The party's position.}

\item{email}{(character) The party's email address(es).}

\item{phone}{(character) The party's phone number(s).}

\item{address}{(character) The party's address(es) as a valid EML address}

\item{userId}{(character) The party's ORCID, in format https://orcid.org/WWWW-XXXX-YYYY-ZZZZ.}

\item{role}{(character) The party's role.}
}
\value{
(party) An instance of the party specified by the \code{type} argument.
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("deprecated")}
}
\details{
Please use the constructors in the EML package instead

You will usually want to use the high-level functions such as
\code{\link[=eml_creator]{eml_creator()}} and \code{\link[=eml_contact]{eml_contact()}} but using this is fine.

The \code{userId} argument assumes an ORCID so be sure to adjust for that.
}
\examples{
\dontrun{
eml_party("creator", "Test", "User")
eml_party("creator", "Bryce", "Mecum", userId = "https://orcid.org/0000-0002-0381-3766")
eml_party("creator", given_names = list("Dominic", "'Dom'"),
                     sur_name = "Mullen", list("NCEAS", "UCSB"),
                     position = list("Data Scientist", "Programmer"),
                         address = eml$address(deliveryPoint = "735 State St",
                         city = "Santa Barbara",
                         administrativeArea = "CA",
                         postalCode = "85719"))
}
}
