% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml.R
\name{eml_project}
\alias{eml_project}
\title{Create an eml-project section.}
\usage{
eml_project(title, awards, first, last, organizations = NULL,
  role = "originator")
}
\arguments{
\item{title}{(character) Title of the project.}

\item{awards}{(character) One or more awards for the project.}

\item{first}{(character) First name of the person with role `role`.}

\item{last}{(character) Last name of the person with role `role`.}

\item{organizations}{(character) Optional. One or more organization strings.}

\item{role}{(character) Optional. Specify an alternate role.}
}
\value{
(project) The new project section.
}
\description{
Note: This is super-limited right now.
}
\examples{
eml_project("Some title", "51231", "Some", "User")
}

