% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml.R
\name{eml_validate_attributes}
\alias{eml_validate_attributes}
\title{Validate an EML attributeList attribute-by-attribute}
\usage{
eml_validate_attributes(attributes)
}
\arguments{
\item{attributes}{(attributeList) An attributeList}
}
\value{
(boolean) Named vector of TRUE/FALSE indicating which attributes
are valid
}
\description{
Validate an EML attributeList attribute-by-attribute
}
\examples{
\dontrun{
atts_df <- read.csv('attributes_table.csv', stringsAsFactors = F)
enum_domain <- read.csv('enumerated_domain.csv') # optional
attributes <- EML::set_attributes(atts_df, factor = enum_domain)
eml_validate_attributes(attributes)
}
}
