% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml.R
\name{eml_validate_full}
\alias{eml_validate_full}
\title{Fully validate an EML document against the KNB's online EML validator}
\usage{
eml_validate_full(eml, quiet = FALSE)
}
\arguments{
\item{eml}{An \code{eml} class object or path to a file.}

\item{quiet}{(logical) Whether or not to print the HTML response. Defaults
to \code{FALSE}.}
}
\value{
The HTML response from the validator as an xml_document.
}
\description{
This function uses the same validation routine that arcticdata.io uses when
you create or update a new science metadata object. The validator that the
\code{\link[EML]{eml_validate}} function uses only checks for
schema validity and produces error messages in another format than
arcticdata.io so it can be hard to know ahead of time whether your upload
will succeed or fail.
}
\details{
You can access this tool in a web browser at
\url{https://knb.ecoinformatics.org/emlparser}.

This function simply shows the HTML response without parsing it into a nice,
readable form so you'll have to read the output to know whether your
document is fully valid.
}
\examples{
\dontrun{
# Get an example EML document from the EML package
eml_path <- system.file(file.path("examples", "example-eml-valid.xml"),
                                  package="EML")
eml_validate_full(eml_path)

# We can also pass in an `eml` object
eml_doc <- EML::read_eml(eml_path)
eml_validate_full(eml_doc)
}
}
