% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{env_get}
\alias{env_get}
\title{environment.R
Author: Bryce Mecum <mecum@nceas.ucsb.edu>}
\usage{
env_get()
}
\value{
(character) The environment name.
}
\description{
Functions related to loading configuriation based upon the environment
the code is being run under.
Get the current environment name.
}
