% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{env_load}
\alias{env_load}
\title{Load environmental variables from a YAML-formatted environment file.}
\usage{
env_load(name = NULL, path = NULL)
}
\arguments{
\item{name}{(character) Optional. The environment name.}

\item{path}{(character) Optional. Path to an environment file.}
}
\value{
(list) A list of name-value pairs.
}
\description{
This file should be formatted in the following way:
}
\details{
some_environment:
  var_one: some value
  var_two: some value
  var_three: some value
}

