% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packaging.R
\name{filter_packaging_statements}
\alias{filter_packaging_statements}
\title{Filter statements related to packaging}
\usage{
filter_packaging_statements(statements)
}
\arguments{
\item{statements}{(data.frame) A set of Statements to be filtered}
}
\value{
(data.frame) The filtered Statements
}
\description{
This is intended to be called after `datapack::getTriples` has been called
on a ResourceMap.
}
\details{
This function was written specifically for the case of updating a resource
map while preserving any extra statements that have been added such as PROV
statements. Statements are filtered according to these rules:
}
