% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_metadata.R
\name{fix_bad_enum}
\alias{fix_bad_enum}
\title{Fix a metadata record with a bad topicCategory.}
\usage{
fix_bad_enum(path)
}
\arguments{
\item{path}{}
}
\value{

}
\description{
This is the case where the ISO schema says what's inside a
gmd:MD_TopicCategoryCode element should match items from a controlled
vocabulary. But in the ISO metadata we have, there are newlines and spaces
around that text which causes a check for string equality to fail. i.e.
}
\details{
'oceans' != '     oceans     '
}

