% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_metadata.R
\name{fix_bad_topic}
\alias{fix_bad_topic}
\title{Fix a metadata record with multiple MD_TopicCategory children elements
inside a single topicCategory element.}
\usage{
fix_bad_topic(path)
}
\arguments{
\item{path}{}
}
\value{

}
\description{
Example:
}
\details{
<gmd:topicCategory>
  <gmd:MD_TopicCategoryCode>imageryBaseMapsEarthCover</gmd:MD_TopicCategoryCode>
  <gmd:MD_TopicCategoryCode>oceans</gmd:MD_TopicCategoryCode>
</gmd:topicCategory>
}

