% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{format_eml}
\alias{format_eml}
\title{Generate the EML 2.1.1 format ID}
\usage{
format_eml(version)
}
\arguments{
\item{version}{The version of EML ('2.1.1' or '2.2.0')}
}
\value{
(character) The format ID for EML 2.1.1.
}
\description{
Returns the EML 2.1.1 format ID.
}
\examples{
format_eml("2.1.1")
\dontrun{
# Upload a local EML 2.1.1 file:
env <- env_load()
publish_object(env$mn, "path_to_some_EML_file", format_eml("2.1"))
}
}
