% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{format_eml_220}
\alias{format_eml_220}
\title{Generate the EML 2.2.0 format ID}
\usage{
format_eml_220()
}
\value{
(character) The format ID for EML 2.2.0
}
\description{
Returns the EML 2.2.0 format ID.
}
\examples{
format_eml()
\dontrun{
# Upload a local EML 2.2.0 file:
env <- env_load()
publish_object(env$mn, "path_to_some_EML_file", format_eml_220())
}
}
