% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{format_iso}
\alias{format_iso}
\title{Generate the ISO 19139 format ID}
\usage{
format_iso()
}
\value{
(character) The format ID for ISO 19139.
}
\description{
Returns the ISO 19139 format ID.
}
\examples{
format_iso()
\dontrun{
# Upload a local ISO19139 XML file:
env <- env_load()
publish_object(env$mn, "path_to_some_EML_file", format_iso())
}
}
