% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packaging.R
\name{generate_resource_map}
\alias{generate_resource_map}
\title{Create a resource map RDF/XML file and save is to a temporary path}
\usage{
generate_resource_map(
  metadata_pid,
  data_pids = NULL,
  child_pids = NULL,
  other_statements = NULL,
  resolve_base = "https://cn.dataone.org/cn/v2/resolve",
  resource_map_pid = NULL
)
}
\arguments{
\item{metadata_pid}{(character) PID of the metadata object.}

\item{data_pids}{(character) PID(s) of the data objects.}

\item{child_pids}{(character) Optional. PID(s) of child resource maps.}

\item{other_statements}{(data.frame) Extra statements to add to the resource map.}

\item{resolve_base}{(character) Optional. The resolve service base URL.}

\item{resource_map_pid}{(character) The PID of a resource map.}
}
\value{
(character) Absolute path to the resource map on disk.
}
\description{
This is a convenience wrapper around the constructor of the \code{ResourceMap}
class from \code{DataPackage}.
}
\examples{
\dontrun{
generate_resource_map("X", "Y", "Z",
                      other_statements = data.frame(subject="http://example.com/me",
                                                    predicate="http://example.com/foo",
                                                    object="http://example.com/bar"))
}
}
