% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packaging.R
\name{generate_resource_map_pid}
\alias{generate_resource_map_pid}
\title{Generate a PID for a new resource map by appending "resource_map_" to it.}
\usage{
generate_resource_map_pid(metadata_pid)
}
\arguments{
\item{metadata_pid}{}
}
\description{
Generate a PID for a new resource map by appending "resource_map_" to it.
}

