% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_all_sysmeta}
\alias{get_all_sysmeta}
\title{Get system metadata for all objects in a data package}
\usage{
get_all_sysmeta(node, pid)
}
\arguments{
\item{node}{(MNode/CNode) The Coordinating/Member Node to run the query on}

\item{pid}{(character) The resource map PID of the package}
}
\value{
(list) A structured list of system metadata
}
\description{
This is a wrapper function around `dataone::getSystemMetadata` that retrieves
the system metadata for all objects in a data package and returns them as a list.
It is useful for inspecting system metadata and identifying changes where needed.
}
\examples{
\dontrun{
# Set environment
cn_staging <- CNode("STAGING2")
knb_test <- getMNode(cn_staging, "urn:node:mnTestKNB")
rm_pid <- "resource_map_urn:uuid:3e5307c4-0bf3-4fd3-939c-112d4d11e8a1"

all <- get_all_sysmeta(knb_test, rm_pid)

# View in RStudio data viewer to inspect
View(all)

# Print specific elements to console
all[["Metadata"]]@rightsHolder

# Create separate object
sysmeta_md <- all[["Metadata"]]
}
}
