% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get_all_versions}
\alias{get_all_versions}
\title{Get the PIDs of all versions of an object.}
\usage{
get_all_versions(node, pid)
}
\arguments{
\item{node}{(MNode|CNode) The node to query.}

\item{pid}{(character) Any object in the chain.}
}
\value{
(character) A vector of PIDs in the chain, in order.
}
\description{
Get the PIDs of all versions of an object.
}

