% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get_chain_neighbors}
\alias{get_chain_neighbors}
\title{Get the obsoleted/obsoletedBy properties of an object as a named list.}
\usage{
get_chain_neighbors(node, pid)
}
\arguments{
\item{node}{(MNode|CNode) The node to query.}

\item{pid}{(character) Any object.}
}
\value{
(list) A list of the obsoleted/obsoletedBy properties.
}
\description{
Get the obsoleted/obsoletedBy properties of an object as a named list.
}
