% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_coord_list}
\alias{get_coord_list}
\title{Get list of Coordinate Reference Systems}
\usage{
get_coord_list()
}
\description{
Get a data.frame of EML coordinate reference systems that can
be searched and filtered more easily than the raw XML file.
}
