% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get_current_version}
\alias{get_current_version}
\title{Get the current package version.}
\usage{
get_current_version()
}
\value{
(character) The current package version.
}
\description{
This function parses the installed DESCRIPTION file to get the latest
version.
}

