% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml.R
\name{get_doc_id}
\alias{get_doc_id}
\title{Get the Metacat docid for the given identifier}
\usage{
get_doc_id(sysmeta)
}
\arguments{
\item{sysmeta}{(SystemMetadata) The sysmeta of the object you want to find.}
}
\value{
(character) The docid
}
\description{
Get the Metacat docid for the given identifier
}
