% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{get_formats}
\alias{get_formats}
\title{Get the list of valid formats from DataONE}
\usage{
get_formats(url = "https://cn.dataone.org/cn/v2/formats")
}
\arguments{
\item{url}{(character) The listFormats endpoint. Defaults to the production
CN}
}
\value{
(character)
}
\description{
Note that this function is intended to return even if the request to the CN
fails. This is so other functions can call continue even if the request
fails.
}
