% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get_identifier}
\alias{get_identifier}
\title{Get the identifier from a DataONE response.}
\usage{
get_identifier(dataone_response)
}
\arguments{
\item{dataone_response}{("XMLInternalDocument" "XMLAbstractDocument")}
}
\value{
(character) The PID.
}
\description{
Example resposne:
}
\details{
<d1:identifier xmlns:d1="http://ns.dataone.org/service/types/v1">
  urn:uuid:12aaf494-5840-434d-9cdb-c2597d58543e
</d1:identifier>
}

