% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get_latest_release}
\alias{get_latest_release}
\title{Use the GitHub API to find the latest release for the package.}
\usage{
get_latest_release()
}
\value{
(character) The latest release.
}
\description{
Use the GitHub API to find the latest release for the package.
}

