% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get_netcdf_format_id}
\alias{get_netcdf_format_id}
\title{Determine the DataONE format ID for the NetCDF file provided by path.}
\usage{
get_netcdf_format_id(path)
}
\arguments{
\item{path}{(character) Full or relative path to the file in question.}
}
\value{
(character) The DataONE format ID.
}
\description{
Determine the DataONE format ID for the NetCDF file provided by path.
}

