% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packaging.R
\name{get_or_create_pid}
\alias{get_or_create_pid}
\title{Get the already-minted PID from the inventory or mint a new one.}
\usage{
get_or_create_pid(file, mn, scheme = "UUID")
}
\arguments{
\item{file}{(data.frame) A single row from the inventory.}

\item{mn}{(MNode) The Member Node that will mint the new PID, if needed.}

\item{scheme}{(character) The identifier scheme to use.}
}
\value{
The identifier (character)
}
\description{
Get the already-minted PID from the inventory or mint a new one.
}

