% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_orcid_email}
\alias{get_orcid_email}
\title{Retrieve an email address from an ORCID URL}
\usage{
get_orcid_email(orcid_url)
}
\arguments{
\item{orcid_url}{(character) A valid ORCID URL address.}
}
\value{
(character) Public e-mail addresses.
}
\description{
Retrieve public email addresses from an ORCID URL.
}
\examples{
\dontrun{
pi_email <- get_orcid_email('https://orcid.org/0000-0002-2561-5840')
}
}
