% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_orcid_name}
\alias{get_orcid_name}
\title{Retrieve a name from an orcid ID URL}
\usage{
get_orcid_name(orcid_url)
}
\arguments{
\item{orcid_url}{(character) A valid orcid ID URL address}
}
\value{
character
}
\description{
Retrieve first and last name from an orcid ID URL by scraping the page.
}
\examples{
\dontrun{
pi_name <- get_orcid_name('https://orcid.org/0000-0002-2561-5840')
}
}
