% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get_package}
\alias{get_package}
\title{Get a structured list of PIDs for the objects in a package.}
\usage{
get_package(mn, pid, parent = FALSE, rows = 1000)
}
\arguments{
\item{mn}{(MNode) The Member Node to run the query on.}

\item{pid}{(character) The the metadata PID of the package.}

\item{parent}{(logical) Whether to query for the parent package. Setting this
to true will make this function take longer to run so it is turned of by
default.}
}
\value{
(list) A structured list of the members of the package.
}
\description{
Get a structured list of PIDs for the objects in a package.
}

