% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get_package_direct}
\alias{get_package_direct}
\title{Get a structured list of PIDs for the objects in a package.}
\usage{
get_package_direct(mn, pid, file_names = FALSE, rows = 1000)
}
\arguments{
\item{mn}{(MNode) The Member Node to run the query on.}

\item{pid}{(character) The the metadata PID of the package.}

\item{file_names}{(logical) Whether to return file names for all objects.}

\item{rows}{(numeric) The number of rows to return in the query. This is only
useful to set if you are warned about the result set being truncated.}
}
\value{

}
\description{
Get a structured list of PIDs for the objects in a package.
}

