% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sysmeta.R
\name{get_related_pids}
\alias{get_related_pids}
\title{Get the related PIDs for a given PID (i.e. data files, resource maps).}
\usage{
get_related_pids(mn, pid)
}
\arguments{
\item{pid}{}
}
\value{
(character) Named character vector.
}
\description{
Get the related PIDs for a given PID (i.e. data files, resource maps).
}

