% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get_resource_map}
\alias{get_resource_map}
\title{Get the resource map(s) for the given object.}
\usage{
get_resource_map(mn, pid, rows = 1000)
}
\arguments{
\item{mn}{(MNode) The Member Node to query.}

\item{pid}{(character) The object to get the resource map(s) for.}

\item{rows}{(numeric) Optional. The number of query results to return. This
shouldn't need to be modified and the default, 1000, is very likely to be
more than enough.}
}
\value{
(character) The resource map(s) that contain `pid`.
}
\description{
Get the resource map(s) for the given object.
}

