% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataone.R
\name{get_token}
\alias{get_token}
\title{Gets the currently set authentication token.}
\usage{
get_token(node)
}
\arguments{
\item{node}{(MNode|CNode) The CN or MN you want to find a token for.}
}
\value{
(character) The token.
}
\description{
Gets the currently set authentication token.
}
\examples{
\dontrun{
cn <- CNode('STAGING2')
mn <- getMNode(cn,"urn:node:mnTestKNB")
get_token(mn)
}
}
