% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get_token_subject}
\alias{get_token_subject}
\title{Returns the subject of the set dataone_test_token}
\usage{
get_token_subject()
}
\value{
(character) The token subject.
}
\description{
Returns the subject of the set dataone_test_token
}

