% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packaging.R
\name{insert_file}
\alias{insert_file}
\title{package.R
Author: Bryce Mecum <mecum@nceas.ucsb.edu>}
\usage{
insert_file(inventory, file, env = NULL)
}
\arguments{
\item{inventory}{(data.frame) An Inventory.}

\item{file}{(character) The fully-qualified relative path to the file. See examples.}

\item{env}{(list) Optional. Specify an environment.}
}
\description{
Code related to inserting datasets as Data Packages.
Insert a file from a single row of the Inventory.
}

