% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packaging.R
\name{insert_package}
\alias{insert_package}
\title{Create a single package Data Package from files in the Inventory.}
\usage{
insert_package(inventory, package, env = NULL)
}
\arguments{
\item{inventory}{(data.frame) An Inventory.}

\item{package}{(character) The package identifier.}

\item{env}{(list) Environment variables.}
}
\value{
A list containing PIDs and whether objects were inserted. (list)
}
\description{
Create a single package Data Package from files in the Inventory.
}
