% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inventory.R
\name{inv_add_extra_columns}
\alias{inv_add_extra_columns}
\title{Adds a set of extra columsn to the inventory that are useful for working
with them.}
\usage{
inv_add_extra_columns(inventory)
}
\arguments{
\item{inventory}{(data.frame) An inventory.}
}
\value{
An inventory (data.frame)
}
\description{
Adds a set of extra columsn to the inventory that are useful for working
with them.
}
