% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inventory.R
\name{inv_add_parent_package_column}
\alias{inv_add_parent_package_column}
\title{Add a column for parent packages.}
\usage{
inv_add_parent_package_column(inventory)
}
\arguments{
\item{inventory}{(data.frame) An Inventory.}
}
\value{
inventory (data.frame) An Inventory.
}
\description{
Add a column for parent packages.
}
