% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inventory.R
\name{inv_init}
\alias{inv_init}
\title{inventory.R
Author: Bryce Mecum <mecum@nceas.ucsb.edu>}
\usage{
inv_init()
}
\value{
An empty data frame
}
\description{
Functions relating to keeping up an inventory of files that exist on the KNB
and may or may not be copied to another computer and untarred.
}
\details{
Create an empty inventory data.frame. This doesn't need to be a function
but I'm making it one in case the initialization routine becomes more
complicated.
}

