% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inventory.R
\name{inv_load_checksums}
\alias{inv_load_checksums}
\title{Load checksums into the inventory file from a text file. This function
removes the column 'checksum_sha256' from inventory before doing a
left join.}
\usage{
inv_load_checksums(inventory, path)
}
\arguments{
\item{inventory}{(data.frame) An inventory.}

\item{path}{(character) Path to a file containing sizes.}
}
\value{
An inventory (data.frame)
}
\description{
Load checksums into the inventory file from a text file. This function
removes the column 'checksum_sha256' from inventory before doing a
left join.
}

