% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inventory.R
\name{inv_load_dois}
\alias{inv_load_dois}
\title{Load DOIs from a text file into the Inventory.}
\usage{
inv_load_dois(inventory, path)
}
\arguments{
\item{inventory}{(data.frame) An inventory.}

\item{path}{Location of a text file with DOIs and file paths. (character)}
}
\value{
(data.frame) The modified Inventory.
}
\description{
Load DOIs from a text file into the Inventory.
}

