% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inventory.R
\name{inv_load_identifiers}
\alias{inv_load_identifiers}
\title{Load identifiers into the inventory file(s) from a text file. This function
removes the column 'identifier' from inventory before doing a
left join.}
\usage{
inv_load_identifiers(inventory, paths)
}
\arguments{
\item{inventory}{(data.frame) An inventory.}

\item{path}{(character) Path(s) to files containing identifiers.}
}
\value{
(data.frame) An inventory.
}
\description{
Load identifiers into the inventory file(s) from a text file. This function
removes the column 'identifier' from inventory before doing a
left join.
}

