% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inventory.R
\name{inv_load_sizes}
\alias{inv_load_sizes}
\title{Load file sizes into an inventory from a text file. Removes the column
'size_bytes' from inventory before doing a left join.}
\usage{
inv_load_sizes(inventory, path)
}
\arguments{
\item{path}{(character) Path to a file containing sizes.}

\item{(data.frame)}{inventory A \code{data.frame}.}
}
\value{
(data.frame) An inventory
}
\description{
Load file sizes into an inventory from a text file. Removes the column
'size_bytes' from inventory before doing a left join.
}

