% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inventory.R
\name{inv_update}
\alias{inv_update}
\title{Update an Inventory with a new Inventory.}
\usage{
inv_update(inventory, new_state)
}
\arguments{
\item{inventory}{(data.frame) The old Inventory.}

\item{new_state}{(data.frame) The new Inventory.}
}
\description{
Update an Inventory with a new Inventory.
}
