% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataone.R
\name{is_authorized}
\alias{is_authorized}
\title{Check if the user has authorization to perform an action on an object.}
\usage{
is_authorized(node, ids, action)
}
\arguments{
\item{node}{(MNode|CNode) The Node to query.}

\item{ids}{(character) The PID or SID to check.}

\item{action}{(character) One of read, write, or changePermission.}
}
\description{
Check if the user has authorization to perform an action on an object.
}

