% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{is_format_id}
\alias{is_format_id}
\title{Test whether an object is a particular format ID.}
\usage{
is_format_id(mn, pid, format_id)
}
\arguments{
\item{mn}{(MNode)}

\item{pid}{(character)}

\item{format_id}{(character)}
}
\value{
(logical)
}
\description{
Test whether an object is a particular format ID.
}

