% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access.R
\name{is_public_read}
\alias{is_public_read}
\title{Check whether an object has public read access}
\usage{
is_public_read(mn, pids, use.names = TRUE)
}
\arguments{
\item{mn}{(MNode) The Member Node.}

\item{pids}{(character) The PIDs of the objects to check for public read access.}

\item{use.names}{(logical) Optional. If `TRUE` (the default), PIDs will
be used as names for the result unless PIDs have names already, in which case
those names will be used for the result.}
}
\value{
(logical) Whether an object has public read access.
}
\description{
Check whether objects have public read access.
No token needs to be set to use this function.
}
\examples{
\dontrun{
cn <- CNode("STAGING2")
mn <- getMNode(cn,"urn:node:mnTestKNB")
pids <- c("urn:uuid:3e5307c4-0bf3-4fd3-939c-112d4d11e8a1",
    "urn:uuid:23c7cae4-0fc8-4241-96bb-aa8ed94d71fe")
is_public_read(mn, pids)
}
}
