% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataone.R
\name{is_token_expired}
\alias{is_token_expired}
\title{Determine whether the set token is expired.}
\usage{
is_token_expired(node)
}
\value{
(boolean)
}
\description{
Determine whether the set token is expired.
}
\examples{
\dontrun{
cn <- CNode('STAGING2')
mn <- getMNode(cn,"urn:node:mnTestKNB")
is_token_expired(mn)
}
}
