% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{log_message}
\alias{log_message}
\title{Log a message to the console and to a logfile.}
\usage{
log_message(message = NULL)
}
\arguments{
\item{message}{(character) Your log message.}
}
\value{
Nothing.
}
\description{
Reads from the environment variable 'LOG_PATH' and uses the value set there
to decide the location of the log file. If that envvar isn't set, it defaults
to 'arcticdata-log.txt'.
}
