% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality.R
\name{mdq_run}
\alias{mdq_run}
\title{Score a metadata document against a MetaDIG suite}
\usage{
mdq_run(document, suite_id = "arctic.data.center.suite.1")
}
\arguments{
\item{document}{(eml/character) Either an EML object or path to a file on disk.}

\item{suite_id}{(character) Specify a suite ID. Should be one of \url{https://quality.nceas.ucsb.edu/quality/suites}.}
}
\value{
(data.frame) A sorted data.frame of check results.
}
\description{
This function scores a metadata document against a MetaDIG suite.
The default suite is for the Arctic Data Center.
}
\examples{
\dontrun{
# Check an EML document you are authoring
library(EML)
mdq_run(new("eml"))

# Check an EML document that is saved to disk
mdq_run(system.file("examples", "example-eml-2.1.1.xml", package = "EML"))
}
}
