% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosaic.R
\name{mosaic_annotate_attribute}
\alias{mosaic_annotate_attribute}
\title{Add an attribute annotation (does not include the id slot)}
\usage{
mosaic_annotate_attribute(eventLabel)
}
\arguments{
\item{eventLabel}{(character) the event ID provided by the researcher}
}
\value{
(list) the annotation
}
\description{
Add an attribute annotation (does not include the id slot)
}
\examples{
mosaic_annotate_attribute("PS122/2_14-270")
}
