% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosaic.R
\name{mosaic_annotate_dataset}
\alias{mosaic_annotate_dataset}
\title{Annotating the dataset level annotations}
\usage{
mosaic_annotate_dataset(campaign)
}
\arguments{
\item{campaign}{(character) the campaign number (can be derrived from the eventID),}
}
\value{
(list) the dataset level annotation
}
\description{
the basis might differ depending on the campaign this function assumes the use of the Polarstern as the basis- please check carefully
}
\examples{
mosaic_annotate_dataset("PS122/2")
}
