% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosaic.R
\name{mosaic_portal_filter}
\alias{mosaic_portal_filter}
\title{Creates the choice label pairs to be pasted into a portal document}
\usage{
mosaic_portal_filter(class)
}
\arguments{
\item{class}{(character) a class in the MOSAiC ontology to get the filters from}
}
\value{
character
}
\description{
The function only selects the annotations that are used for method/devices (there are 500 + options).
copy and paste the output into a portal document's choice filters
}
\examples{
mosaic_portal_filter("Method/Device")

mosaic_portal_filter("Basis")

mosaic_portal_filter("Campaign")
}
