% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{object_exists}
\alias{object_exists}
\title{Check if an object exists on a Member Node.}
\usage{
object_exists(node, pids)
}
\arguments{
\item{node}{(MNode|CNode) The Node to query.}

\item{pid}{(character) PID to check the existence of.}
}
\value{
(logical) Whether the object exists.
}
\description{
This is a simple check for the HTTP status of a /meta/{PID} call on the
provided member node.
}
