% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packaging.R
\name{parse_resource_map}
\alias{parse_resource_map}
\title{Parse a resource map into a data.frame}
\usage{
parse_resource_map(path)
}
\arguments{
\item{path}{(character) Path to the resource map (an RDF/XML file).}
}
\value{
(data.frame) The statements in the resource map.
}
\description{
Parse a resource map into a data.frame.
}
\examples{
\dontrun{
# Set environment
cn <- CNode("STAGING2")
mn <- getMNode(cn,"urn:node:mnTestKNB")

rm_pid <- "resource_map_urn:uuid:6b2e5753-4a94-4e6f-971c-36420a446ecb"

# Write resource map to file
writeBin(getObject(mn, rm_pid), "~/Documents/resource_map.rdf")
df <- parse_resource_map("~/Documents/resource_map.rdf")
}
}
