% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{path_join}
\alias{path_join}
\title{(Intelligently) join (possibly redudant) path parts together.}
\usage{
path_join(path_parts = c(""))
}
\arguments{
\item{path_parts}{(character)}
}
\value{
(character)The joined path string.
}
\description{
Joins path strings like "./" to "./my/dir" as "./my/dir" instead of as
"././my/dir.
}

