% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml.R
\name{pid_to_eml_datatable}
\alias{pid_to_eml_datatable}
\title{This function is deprecated. See \link{pid_to_eml_entity}.}
\usage{
pid_to_eml_datatable(mn, pid, attributes = NULL, factors = NULL,
  name = NULL, description = NULL, validateAttributes = TRUE)
}
\arguments{
\item{mn}{(MNode) Member Node where the PID is associated with an object.}

\item{pid}{(character) The PID of the object to create the \code{dataTable} for.}

\item{attributes}{(data.frame) Optional data frame of attributes. Follows the convention in \link[EML]{set_attributes}.}

\item{factors}{(data.frame) Optional data frame of enumerated attribute values (factors). Follows the convention in \link[EML]{set_attributes}.}

\item{name}{(character) Optional field to specify \code{entityName}, otherwise will be extracted from system metadata.}

\item{description}{(character) Optional field to specify \code{entityDescription}, otherwise will match name.}

\item{validateAttributes}{(logical) If set to FALSE or if attributes are not passed into the function, attribute validatio}
}
\description{
This function is deprecated. See \link{pid_to_eml_entity}.
}
