% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml.R
\name{pid_to_eml_other_entity}
\alias{pid_to_eml_other_entity}
\title{This function is deprecated. See \link{pid_to_eml_entity}.}
\usage{
pid_to_eml_other_entity(mn, pids)
}
\arguments{
\item{mn}{(MNode) Member Node where the PID is associated with an object.}

\item{pids}{(character) The PID of the object to create the sub-tree for.}
}
\description{
This function is deprecated. See \link{pid_to_eml_entity}.
}
