% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml.R
\name{pid_to_eml_physical}
\alias{pid_to_eml_physical}
\title{Create EML physical objects for the given set of PIDs}
\usage{
pid_to_eml_physical(mn, pid)
}
\arguments{
\item{mn}{(MNode) Member Node where the PID is associated with an object.}

\item{pid}{(character) The PID of the object to create the physical for.}
}
\value{
(list) A physical object.
}
\description{
This is a wrapper around \code{\link[=sysmeta_to_eml_physical]{sysmeta_to_eml_physical()}} which handles the task of
creating the EML physical.
}
\examples{
\dontrun{
# Generate EML physical sections for an object in a data package
phys <- pid_to_eml_physical(mn, pid)
}
}
