% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml.R
\name{pid_to_entity}
\alias{pid_to_entity}
\title{eml.R}
\usage{
pid_to_entity(mn, pid, sysmeta = NULL)
}
\arguments{
\item{mn}{(MNode) Member Node where the PID is associated with an object.}

\item{pid}{(character) The PID of the object to create the sub-tree for.}

\item{sysmeta}{(SystemMetadata) Optional. Manually pass in System Metadata. This avoids an extra network request if the calling environment has already loaded the System Metadata.}
}
\value{
(otherEntity) The XML sub-tree.
}
\description{
Helpers for creating EML.
Create an EML otherEntity sub-tree for the given PID.
}
\details{
Note this is a wrapper around sysmeta_to_entity which handles the task of
creating the EML otherEntity sub-ree.
}

