% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_metadata.R
\name{pretty_print}
\alias{pretty_print}
\title{Uses XMLStarlet to pretty-print/beautify an XML document.}
\usage{
pretty_print(path)
}
\arguments{
\item{path}{Path to your file you want pretty-printed. (character)}
}
\value{
Returns the result of the `system` command (0 = success)
}
\description{
This command just runs `xmlstarlet path > path`, doing a simple
pretty-printing of the file located at `path`.
}
\details{
Note that this function is doing an in-place pretty printing instead of
returning the pretty-printed text.

Note that this command uses a temporary file as an intermediate step in the
pretty-printing process. For some reason, when running xmlstarlet from within
R, the same file can't be used as the input to `xmlstarlet format` and as the
shell redirection file (`> somefile.txt`). If you try to run `xmlstarlet
format` on the same file as you redirect to, you get a weird parse error from
xmlstarlet.
}

