% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editing.R
\name{publish_update}
\alias{publish_update}
\title{Publish an updated data package.}
\usage{
publish_update(mn, metadata_old_pid, resmap_old_pid, data_old_pids = NULL,
  child_pids = NULL, metadata_file_path = NULL, use_doi = FALSE,
  parent_resmap_pid = NULL, parent_metadata_pid = NULL,
  parent_data_pids = NULL, parent_child_pids = NULL)
}
\arguments{
\item{mn}{(MNode)The Member Node to update the object on.}

\item{metadata_old_pid}{The PID of the EML metadata document to be updated}

\item{resmap_old_pid}{The PID of the resource map for the package}

\item{data_old_pids}{a vector of PIDs of data objects in the package}

\item{use_doi}{boolean indicating if a DOI should be used for the metadata}

\item{parent_resmap_pid}{optional PID of a parent package to be updated}

\item{parent_metadata_pid}{optional identifier for the metadata document of the parent package}

\item{parent_data_pids}{optional vector of identifiers of data in the parent package}

\item{parent_child_pids}{optional vector of identifiers of child packages in the parent package}

\item{metadata_file}{optional filename  of a replacement EML file for the metadata}
}
\description{
The metadata_old_pid and resmap_old_pid provide the identifier of an EML metadata
document and associated resource map, and the data_old_pids vector provides a list
of PIDs of data objects in the package.  Update the metadata file and resource map
by generating a new identifier (a DOI if use_doi is TRUE) and updating the Member
Node with a public version of the object.  If metadata_file is not missing, it
should be an edited version of the metadata to be used to update the original. If
parent_resmap_pid is not missing, it indicates the PID of a parent package that
should be updated as well, using the parent_medata_pid, parent_data_pids, and
parent_child_pids as members of the updated package. In all cases, the objects
are made publicly readable.
}

