% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology.R
\name{read_ontology}
\alias{read_ontology}
\title{Get an owl file from github}
\usage{
read_ontology(ontology_name)
}
\arguments{
\item{ontology_name}{the name of the onotology to read; one of mosaic or ecso}
}
\value{
list
}
\description{
Get an owl file from github
}
\examples{
read_ontology("mosaic")
read_ontology("ecso")
}
