% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{read_zip_shapefile}
\alias{read_zip_shapefile}
\title{Read a shapefile from a pid}
\usage{
read_zip_shapefile(mn, pid)
}
\arguments{
\item{mn}{(MNode) A DataOne Member Node}

\item{pid}{(character) An object identifier}
}
\value{
shapefile (sf) The shapefile as an \code{sf} object
}
\description{
Read a shapefile from a pid that points to the zipped directory of the shapefile and associated files
on a given member node.
}
\examples{
\dontrun{
cn <- dataone::CNode('PROD')
adc <- dataone::getMNode(cn,'urn:node:ARCTIC')
pid <- "urn:uuid:294a365f-c0d1-4cc3-a508-2e16260aa70c"

shapefile <- read_zip_shapefile(adc, pid)
}
}
\author{
Jeanette Clark jclark@nceas.ucsb.edu
}
