% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{recover_failed_submission}
\alias{recover_failed_submission}
\title{Recovers failed submissions}
\usage{
recover_failed_submission(node, pid, path)
}
\arguments{
\item{node}{(MNode) The Member Node to publish the object to.}

\item{pid}{The PID of the EML metadata document to be recovered.}

\item{path}{path to write XML.}
}
\value{
recovers and write the valid EML to the indicated path
}
\description{
Recovers failed submissions and write the new, valid EML to a given path
}
\examples{
\dontrun{
# Set environment
cn <- dataone::CNode("STAGING2")
mn <- dataone::getMNode(cn,"urn:node:mnTestKNB")
pid <- "urn:uuid:c40b93d9-a15a-47d0-9a5f-06e7056c93c1"
path <- tempfile("file", fileext = ".xml")
recover_failed_submission(mn, pid, path)
eml <- EML::read_eml(path)
}
}
\author{
Rachel Sun rachelsun@ucsb.edu
}
