% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editing.R
\name{recover_prov}
\alias{recover_prov}
\title{Get a data.frame of prov statements from a resource map pid.}
\usage{
recover_prov(mn, rm_pid)
}
\arguments{
\item{mn}{(mn) A memeber node instance}

\item{rm_pid}{(character) A resource map identifier}
}
\value{
a data.frame of prov statments
}
\description{
This is a function that is useful if you need to recover lost prov statements. It returns
a data.frame of statements that can be passed to \code{update_resource_map} in the \code{other_statements}
argument.
}
