% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editing.R
\name{reformat_file_name}
\alias{reformat_file_name}
\title{Helper for publish_object. Reformat the fileName in system metadata.}
\usage{
reformat_file_name(path, sysmeta)
}
\arguments{
\item{path}{(character) full file path}

\item{sysmeta}{(S4) A system metadata object}
}
\description{
Reformat the fileName field in an object's system metadata to follow Arctic Data Center
system metdata naming conventions.  Publish_object calls this function to rename
the fileName field in system metadata.
}
