% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sysmeta.R
\name{remove_public_access}
\alias{remove_public_access}
\title{Remove all public read access rules from a System Metadata document}
\usage{
remove_public_access(sysmeta)
}
\arguments{
\item{sysmeta}{(SystemMetadata) The System Metadata document. to change.}
}
\value{
(SystemMetadata) The potentially modified System Metadata document.
}
\description{
Remove all public read access rules from a System Metadata document
}
\examples{
library(datapack)
sm <- new("SystemMetadata)
sm <- addAccessRule(sm, "public", "read)
sm@accessPolicy
sm <- remove_public_access(sm)
sm@accessPolicy
}

