% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml.R
\name{reorder_pids}
\alias{reorder_pids}
\title{Reorder a named list of objects according to the order in the metadata}
\usage{
reorder_pids(pid_list, doc)
}
\arguments{
\item{pid_list}{(list) A named list of data pids}

\item{doc}{(list) an \code{emld} document}
}
\value{
ordered_pids (list) A list of reordered pids
}
\description{
This function takes a named list of data objects, such as what is
returned from \code{get_package}, and reorders them according to the order
they are given in the EML document.
}
\examples{
\dontrun{
cn <- dataone::CNode('PROD')
adc <- dataone::getMNode(cn,'urn:node:ARCTIC')
ids <- get_package(adc, 'resource_map_doi:10.18739/A2S17SS1M', file_names = TRUE)
doc <- EML::read_eml(dataone::getObject(adc, ids$metadata))

# return all entity types
ordered_pids <- reorder_pids(ids$data, doc)
}

}
