% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{replace_package_id}
\alias{replace_package_id}
\title{Replace the EML 'packageId' attribute on the root element with a
certain value.}
\usage{
replace_package_id(path, replacement)
}
\arguments{
\item{path}{(character) Path to the XML file to edit.}

\item{replacement}{(character) The new value.}
}
\description{
Replace the EML 'packageId' attribute on the root element with a
certain value.
}
