% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sysmeta.R
\name{replace_subject}
\alias{replace_subject}
\title{Replace subjects in the accessPolicy section of a System Metadata entries.}
\usage{
replace_subject(sysmeta,
  from = "cn=arctic-data-admins,dc=dataone,dc=org",
  to = "CN=arctic-data-admins,DC=dataone,DC=org")
}
\arguments{
\item{sysmeta}{(SystemMetadata) The System Metadata object.}

\item{from}{(character) The DN string to replace.}

\item{to}{(character) The DN string to put in place of `from`.}
}
\value{
The modified System Metadata (SystemMetadata)
}
\description{
This function was written out to fix capitalization errors but in a set of
existing System Metadata entries but can be used to replace any subject.
}
