% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editing.R
\name{set_file_name}
\alias{set_file_name}
\title{Set the file name for an object}
\usage{
set_file_name(mn, pid, name)
}
\arguments{
\item{mn}{(MNode) The Member Node.}

\item{pid}{(character) The PID of the object to set the file name on.}

\item{name}{(character) The file name.}
}
\value{
(logical) Whether the update succeeded.
}
\description{
Set the file name for an object.
}
\examples{
\dontrun{
cn <- CNode("STAGING2")
mn <- getMNode(cn, "urn:node:mnTestKNB")

pid <- "urn:uuid:23c7cae4-0fc8-4241-96bb-aa8ed94d71fe"
set_file_name(mn, pid, "myfile.csv")
}
}
