% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml.R
\name{set_other_entities}
\alias{set_other_entities}
\title{This function is deprecated. See \link{pid_to_eml_entity}.}
\usage{
set_other_entities(mn, path, pids)
}
\arguments{
\item{mn}{(MNode) The Member Node the objects exist on.}

\item{path}{(character) The location on disk of the EML file.}

\item{pids}{(character) One or more PIDs for the objects.}
}
\description{
This function is deprecated. See \link{pid_to_eml_entity}.
}
