% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{set_public_read_all_versions}
\alias{set_public_read_all_versions}
\title{Set public READ access on all versions of PIDs in data package.}
\usage{
set_public_read_all_versions(mn, resource_map_pid)
}
\arguments{
\item{mn}{(MNode) The Member Node to query.}

\item{resource_map_pid}{(character) The resource map identifier (PID).}
}
\description{
Set public READ access on all versions of PIDs in data package.
}
\examples{
\dontrun{
cn_staging <- CNode('STAGING')
adc_test <- getMNode(cn_staging,'urn:node:mnTestARCTIC')
# Create a dummy package then create another version with 'publish_update()'
pkg <- create_dummy_package(adc_test)
remove_public_read(mn, unlist(pkg))
pkg_v2 <- publish_update(adc_test, pkg$metadata, pkg$resource_map, pkg$data, public = FALSE)
# Set public read on all versions
set_public_read_all_versions(adc_test, pkg$resource_map)
}
}
