% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{show_indexing_status}
\alias{show_indexing_status}
\title{Show the indexing status of a set of PIDs}
\usage{
show_indexing_status(mn, pids)
}
\arguments{
\item{mn}{(MNode) The Member Node to query}

\item{pids}{(character|list) One or more PIDs (or list of PIDs)}
}
\value{
Nothing
}
\description{
Show the indexing status of a set of PIDs
}
\examples{
\dontrun{
# Create a package then check its indexing status
library(dataone)
mn <- MNode(...)
pkg <- create_dummy_package(mn)
show_indexing_status(mn, pkg)
}
}
