% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{show_random_dataset}
\alias{show_random_dataset}
\title{Print a random dataset.}
\usage{
show_random_dataset(inventory, theme = NULL, n = 10)
}
\arguments{
\item{inventory}{(data.frame) An inventory.}

\item{theme}{(character) Optional. A package theme name.}

\item{n}{(numeric) Optional. The number of files to show.}
}
\value{
Nothing.
}
\description{
Print a random dataset.
}
