% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{substitute_eml_party}
\alias{substitute_eml_party}
\title{Extract the EML responsible-party blocks in a document, and parse the
surName field to create proper givenName/surName structure}
\usage{
substitute_eml_party(path)
}
\arguments{
\item{path}{file path to the EML document to process (character)}
}
\value{
path (character) Path to the converted EML file.
}
\description{
Extract the EML responsible-party blocks in a document, and parse the
surName field to create proper givenName/surName structure
}

