% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_metadata.R
\name{test_has_abstract}
\alias{test_has_abstract}
\title{modify_metadata.R
Author: Bryce Mecum <mecum@nceas.ucsb.edu>}
\usage{
test_has_abstract(path)
}
\description{
Functions related to fixing invalid ISO metadata.
}
\details{
Some functions just test whether a validation issue is present. These are
prefixed with the text "test". Exactly what they are testing should be
described in the docstrings.

Other functons fix the bad metadata in place (modifying the original file)
and these functions are prefixed with "fix_". Exactly what they are fixing
should be described in the docstrings.

Example usage:

# Find and fix documents in 'mydir' that have extra whitespace in their
# topicCategory element(s)

the_files <- dir(mydir)
bad_enums <- the_files[which(sapply(the_files, test_has_bad_enum))]
}
