% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marking.R
\name{theme_packages}
\alias{theme_packages}
\title{marking.R
Author: Bryce Mecum <mecum@nceas.ucsb.edu>}
\usage{
theme_packages(inventory, nfiles_cutoff = 100)
}
\arguments{
\item{inventory}{(data.frame) An Inventory.}
}
\value{
(data.frame) An Inventory.
}
\description{
R commands for marking datasets before adding.
Divide packages and their files into themes.
}
\details{
Themes divide packages into groups based upon how the actions we will take
to insert them. Packages are divided into one of three themes:

"many-files"

The package has more files than we'd like to include in a Resource Map
   and we will want to archive its contents before inserting.

"has-versions"

The package has version information embedded into its folder structure.
   These packages will be hand-verified and inserted manually when a plan
   is developed to insert them.

"ready-to-go"

All other packages not in the above themes.

Note: Adds a 'theme' column to 'inventory'.
Note: Depeneds on the following columns:

- filename
 - package_nfiles
}
