% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editing.R
\name{title_to_file_name}
\alias{title_to_file_name}
\title{Formats the eml file name based on the dataset title}
\usage{
title_to_file_name(title)
}
\arguments{
\item{title}{(character) title of the dataset}
}
\value{
(character) file path with underscores and extension (.xml)
}
\description{
Formats the eml file name based on the dataset title
}
\examples{
title_to_file_name("Example title here")
}
