% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packaging.R
\name{update_package}
\alias{update_package}
\title{Update a package with modified metadata.}
\usage{
update_package(inventory, package, env = NULL)
}
\arguments{
\item{inventory}{(data.frame)}

\item{package}{(character)}
}
\value{
TRUE or FALSE depending on sucess (logical)
}
\description{
The modified metadata should be set in the `env` variable. For example, if
your original metadata is:
}
\details{
/home/you/originals/dir/a.xml

and your modified metadata is in

/home/someone_else/modified/dir/a.xml

Then your env should be:

env$base_path <- "/home/you/"
env$alternate_path <- "/home/someone_else"

Note that the data files are not updated either so all that's happening is
the metadata object and resource map are being updated.

Note that this function checks if the old objects (metadata and resource map)
exist on the Member Node before doing their work and will call createObject()
instead of updateObject() if the object didn't already exist.
}
