% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editing.R
\name{update_physical}
\alias{update_physical}
\title{Update physical of an updated data object}
\usage{
update_physical(eml, mn, data_pid, new_data_pid)
}
\arguments{
\item{eml}{(eml) An EML class object.}

\item{mn}{(MNode) The Member Node of the data package.}

\item{data_pid}{(character) The identifier of the data object to be updated.}

\item{new_data_pid}{(character) The new identifier of the updated data object.}
}
\description{
This function updates the EML with the new physical
of a data object once it has been updated.
This is a helper function for \code{\link{update_package_object}}.
}
