% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive.R
\name{view_packages}
\alias{view_packages}
\title{interactive.R
Author: Bryce Mecum <mecum@nceas.ucsb.edu>}
\usage{
view_packages(inventory)
}
\description{
Functions for interactive viewing of the Inventory and other objects.
}

