% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{warn_current_version}
\alias{warn_current_version}
\title{Warns if the currently-installed version of the package is not the same
version as the latest release on GitHub.}
\usage{
warn_current_version()
}
\description{
Warns if the currently-installed version of the package is not the same
version as the latest release on GitHub.
}

