% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcticdatautils.R
\docType{package}
\name{arcticdatautils}
\alias{arcticdatautils}
\alias{arcticdatautils-package}
\title{arcticdatautils: Utilities for the Arctic Data Center}
\description{
The foo package provides three categories of important functions:
foo, bar and baz.
}
\section{Foo functions}{

The foo functions ...
}

