% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcticdatautils.R
\docType{package}
\name{arcticdatautils}
\alias{arcticdatautils}
\title{arcticdatautils: Utilities for the Arctic Data Center}
\description{
This package contains code for doing lots of useful stuff that's too specific for the
dataone package, primarily functions that streamline Arctic Data Center operations.
}
