% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml.R
\name{clear_methods}
\alias{clear_methods}
\title{Clear all methods}
\usage{
clear_methods(doc)
}
\arguments{
\item{doc}{(eml) The document to clear methods from.}
}
\value{
(eml) The modified EML document.
}
\description{
Clear all methods from an EML document.
}
\examples{
\dontrun{
eml <- read_eml("~/Documents/metadata.xml")
eml <- clear_methods(eml)
}
}
