% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{convert_iso_to_eml}
\alias{convert_iso_to_eml}
\title{Convert and ISO document to EML using an XSLT.}
\usage{
convert_iso_to_eml(path, style = NA)
}
\arguments{
\item{path}{(character) Path to the file to convert.}

\item{style}{(xslt) The XSLT object to be used for transformation.}
}
\value{
(character) Location of the converted file.
}
\description{
Leave style=NA if you want to use the default ISO-to-EML stylesheet.
}
