% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{create_dummy_enumeratedDomain_dataframe}
\alias{create_dummy_enumeratedDomain_dataframe}
\title{Create test enumeratedDomain data.frame}
\usage{
create_dummy_enumeratedDomain_dataframe(factors)
}
\arguments{
\item{factors}{(character) Vector of factor names to include.}
}
\value{
(data.frame) A data.frame of factors.
}
\description{
Create a test data.frame of enumeratedDomains.
}
\examples{
\dontrun{
# Create dummy dataframe of 2 factors/enumerated domains
attributes <- create_dummy_enumeratedDomain_dataframe(c("Factor1", "Factor2"))
}
}
