% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{create_dummy_enumeratedDomain_dataframe}
\alias{create_dummy_enumeratedDomain_dataframe}
\title{Create dummy enumeratedDomain data frame}
\usage{
create_dummy_enumeratedDomain_dataframe(factors)
}
\arguments{
\item{factors}{(character) Vector of factor names to include.}
}
\value{
(data.frame) Data frame of factors
}
\description{
Create dummy enumeratedDomain data frame
}
\examples{
\dontrun{
# Create dummy dataframe of 2 factors/enumerated domains
attributes <- create_dummy_enumeratedDomain_dataframe(c("Factor1", "Factor2"))
}
}
