% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{create_dummy_metadata}
\alias{create_dummy_metadata}
\title{helpers.R}
\usage{
create_dummy_metadata(mn, data_pids = NULL)
}
\arguments{
\item{mn}{(MNode) The Member Node.}

\item{data_pids}{(character) Optional. PIDs for data objects the metadata documents.}
}
\description{
Various helper functions for things like testing the package.
Create a test metadata object.
}

