% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{create_dummy_package}
\alias{create_dummy_package}
\title{Create a test package}
\usage{
create_dummy_package(mn, size = 2)
}
\arguments{
\item{mn}{(MNode) The Member Node.}

\item{size}{(numeric) The number of files in the package, including the metadata file.}
}
\value{
(character) A named character vector of the data PIDs in the package.
}
\description{
Create a test data package.
}
\examples{
\dontrun{
# Set environment
cn <- CNode("STAGING2")
mn <- getMNode(cn,"urn:node:mnTestKNB")
#Create dummy package with 5 data objects and 1 metadata object
pids <- create_dummy_package(mn, 6)
}
}
