% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{create_dummy_package}
\alias{create_dummy_package}
\title{Create a test package.}
\usage{
create_dummy_package(mn, size = 2)
}
\arguments{
\item{mn}{(MNode) The Member Node.}

\item{size}{(numeric) The number of files in the package.}
}
\description{
Create a test package.
}
