% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editing.R
\name{create_resource_map}
\alias{create_resource_map}
\title{Create a resource map.}
\usage{
create_resource_map(mn, metadata_pid, data_pids = NULL, child_pids = NULL,
  check_first = TRUE)
}
\arguments{
\item{check_first}{(logical) Optional. Whether to check the PIDs passed in as aruments exist on the MN before continuing. This speeds up the function, especially when `data_pids` has many elements.}
}
\value{
The created resource map's PID (character)
}
\description{
Similar to but different than update_resource_map in that it uses a create
call instead of an update call.
}

