% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml.R
\name{eml_abstract}
\alias{eml_abstract}
\title{Minimalistic helper function to generate EML abstracts}
\usage{
eml_abstract(text)
}
\arguments{
\item{text}{(character) Paragraphs of text, one paragraph per element in the
character vector}
}
\value{
(abstract) An EML abstract
}
\description{
Minimalistic helper function to generate EML abstracts
}
\examples{
# Set an abstract with a single paragraph
eml_abstract("Test abstract...")

# Or one with multiple paragraphs
eml_abstract(c("First para...", "second para..."))
}
