% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml.R
\name{eml_abstract}
\alias{eml_abstract}
\title{Create an EML abstract}
\usage{
eml_abstract(text)
}
\arguments{
\item{text}{(character) Paragraphs of text with one paragraph per element in the
character vector, constructed using \code{list}}
}
\value{
(abstract) An EML abstract.
}
\description{
Create an EML abstract.
}
\details{
Note that eml$abstract() provides the same functionality.
}
\examples{
# Set an abstract with a single paragraph
eml_abstract("Test abstract...")

# Or one with multiple paragraphs
eml_abstract(list("First para...", "second para..."))
}
