% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml.R
\name{eml_associated_party}
\alias{eml_associated_party}
\title{Create an EML associatedParty}
\usage{
eml_associated_party(...)
}
\arguments{
\item{...}{Arguments passed on to eml_party}
}
\value{
(associatedParty) The new associatedParty
}
\description{
See \code{\link{eml_party}} for details.
}
\examples{
eml_associated_party("test", "user", email = "test@user.com", role = "Principal Investigator")
}
