% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml.R
\name{eml_contact}
\alias{eml_contact}
\title{Create an EML contact subtree from a first, last, and email.}
\usage{
eml_contact(first, last, email)
}
\arguments{
\item{first}{(character) The first name.}

\item{last}{(character) The last name.}

\item{email}{(character) The email address.}
}
\value{
(contact) The new contact sub-tree.
}
\description{
Create an EML contact subtree from a first, last, and email.
}

