% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml.R
\name{eml_contact}
\alias{eml_contact}
\title{Create an EML contact}
\usage{
eml_contact(...)
}
\arguments{
\item{...}{Arguments passed on to eml_party}
}
\value{
(contact) The new contact
}
\description{
See \code{\link{eml_party}} for details.
}
\examples{
eml_contact("test", "user", email = "test@user.com")
}
