% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml.R
\name{eml_creator}
\alias{eml_creator}
\title{Create an EML creator}
\usage{
eml_creator(given_names, sur_name, organization = NULL, email = NULL,
  phone = NULL, address = NULL)
}
\arguments{
\item{given_names}{(character) One or more given (first) names.}

\item{sur_name}{(character) A sur (last) name.}

\item{organization}{(character) One or more organization names.}

\item{email}{(character) An email address.}

\item{phone}{(character) A phone number.}

\item{address}{(address) An object of type 'address' (EML).}
}
\value{
(creator) The new creator
}
\description{
Create an EML creator
}

