% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml.R
\name{eml_creator}
\alias{eml_creator}
\title{Create an EML creator subtree from a first, last, and email.}
\usage{
eml_creator(first, last, email)
}
\arguments{
\item{first}{(character) The first name.}

\item{last}{(character) The last name.}

\item{email}{(character) The email address.}
}
\value{
(creator) The new creator sub-tree.
}
\description{
Create an EML creator subtree from a first, last, and email.
}

