% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{eml_creator}
\alias{eml_creator}
\title{Create an EML creator.}
\usage{
eml_creator(...)
}
\arguments{
\item{...}{Arguments passed on to \code{\link[=eml_party]{eml_party()}}.}
}
\value{
(creator) The new creator.
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("deprecated")}
}
\details{
Please use the constructors in the EML package instead
}
\examples{
\dontrun{
eml_creator("test", "user", email = "test@user.com")
eml_creator("creator", "Bryce", "Mecum", userId = "https://orcid.org/0000-0002-0381-3766")
eml_creator("creator", c("Dominic", "'Dom'"), "Mullen", c("NCEAS", "UCSB"),
            c("Data Scientist", "Programmer"))
}
}
