% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml.R
\name{eml_individual_name}
\alias{eml_individual_name}
\title{Create an EML individualName section}
\usage{
eml_individual_name(given_names = NULL, sur_name)
}
\arguments{
\item{given_names}{(character) One or more given names.}

\item{sur_name}{(character) A sur (last) name.}
}
\value{
(individualName) The new individualName section
}
\description{
Create an EML individualName section
}
