% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml.R
\name{eml_metadata_provider}
\alias{eml_metadata_provider}
\title{Create an EML metadataProvider}
\usage{
eml_metadata_provider(...)
}
\arguments{
\item{...}{Arguments passed on to \code{\link[=eml_party]{eml_party()}}.}
}
\value{
(metadataProvider) The new metadataProvider.
}
\description{
See \code{\link[=eml_party]{eml_party()}} for details.
}
\examples{
eml_metadata_provider("test", "user", email = "test@user.com")
}
