% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml.R
\name{eml_otherEntity_to_dataTable}
\alias{eml_otherEntity_to_dataTable}
\title{Convert otherEntities to dataTables}
\usage{
eml_otherEntity_to_dataTable(eml, otherEntity, validate_eml = TRUE)
}
\arguments{
\item{eml}{(S4) An EML S4 object}

\item{otherEntity}{(S4 / integer) Either an EML otherEntity object or the index
of an otherEntity within a ListOfotherEntity.  Integer input is recommended.}

\item{validate_eml}{(logical) Optional.  Specify whether or not to validate the eml after
completion.  Defaults to \code{TRUE}.  Recommended setting is \code{TRUE}.  Setting this to
\code{FALSE} reduces execution time by ~50%.}
}
\description{
Convert an EML 'otherEntity' object to a 'dataTable' object.  It converts the
otherEntity as currently constructed - it does not add a physical or add attributes.
However, if these are already in their respective slots, they will be retained.
}
\examples{
\dontrun{
eml <- read_eml(system.file("example-eml.xml", package = "arcticdatautils"))

# The following two calls are equivalent:
eml <- eml_otherEntity_to_dataTable(eml, eml@dataset@otherEntity[[1]])
eml <- eml_otherEntity_to_dataTable(eml, 1)

# Integer input is recommended:
eml <- eml_otherEntity_to_dataTable(eml, 1)
}
}
\author{
Dominic Mullen dmullen17@gmail.com
}
