% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml.R
\name{eml_otherEntity_to_dataTable}
\alias{eml_otherEntity_to_dataTable}
\title{Convert otherEntities to dataTables}
\usage{
eml_otherEntity_to_dataTable(doc, index, validate_eml = TRUE)
}
\arguments{
\item{doc}{(list) An EML document.}

\item{validate_eml}{(logical) Optional. Whether or not to validate the EML after
completion. Setting this to \code{FALSE} reduces execution time by ~50 percent.}

\item{otherEntity}{(integer/) The indicies of the otherEntities to be transformed.}
}
\description{
Convert an EML 'otherEntity' object to a 'dataTable' object. This will convert an
otherEntity object as currently constructed - it does not add a physical or add attributes.
However, if these are already in their respective slots, they will be retained.
}
\examples{
\dontrun{
doc <- read_eml(system.file("example-eml.xml", package = "arcticdatautils"))

# The following two calls are equivalent:
doc <- eml_otherEntity_to_dataTable(doc, doc$dataset$otherEntity[[1]])
doc <- eml_otherEntity_to_dataTable(doc, 1)

# Integer input is recommended:
doc <- eml_otherEntity_to_dataTable(doc, 1)
}
}
\author{
Dominic Mullen dmullen17@gmail.com
}
