% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml.R
\name{eml_party}
\alias{eml_party}
\title{Low-level helper for creating EML parties}
\usage{
eml_party(type = "associatedParty", given_names = NULL, sur_name = NULL,
  organization = NULL, position = NULL, email = NULL, phone = NULL,
  address = NULL, userId = NULL, role = NULL)
}
\arguments{
\item{type}{(character) The type of party (e.g. 'contact')}

\item{given_names}{(character) The party's given name(s)}

\item{sur_name}{(character) The party's surname}

\item{organization}{(character) The party's organization name}

\item{position}{(character) The party's position}

\item{email}{(character) The party's email address(es)}

\item{phone}{(character) The party's phone number(s)}

\item{address}{(character) The party's address(es)}

\item{userId}{(character) The party's ORCID, in format https://orcid.org/WWWW-XXXX-YYYY-ZZZZ}

\item{role}{(character) The party's role}
}
\value{
An instance of the party specified by the in \code{type} argument
}
\description{
You usually will want to use the high-level functions such as
\code{\link{eml_creator}} and \code{\link{eml_contact}} but using this is
fine.
}
\details{
The \code{userId} argument assumes an ORCID so be sure to adjust for that.
}
