% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml.R
\name{eml_party}
\alias{eml_party}
\title{Low-level helper for creating EML parties}
\usage{
eml_party(type = "associatedParty", given_names = NULL, sur_name = NULL,
  organization = NULL, position = NULL, email = NULL, phone = NULL,
  address = NULL, role = NULL)
}
\arguments{
\item{role}{}
}
\value{
The party set in \code{type}
}
\description{
You usually will want to use the high-level functions such as
\code{\link{eml_creator}} and \code{\link{eml_contact}} but using this is
fine.
}
