% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml.R
\name{eml_project}
\alias{eml_project}
\title{Create an eml-project section.}
\usage{
eml_project(title, personnelList, abstract = NULL, funding = NULL,
  studyAreaDescription = NULL, designDescription = NULL,
  relatedProject = NULL)
}
\arguments{
\item{title}{(character) Title of the project (Required).}

\item{personnelList}{(list of personnel) Personnel involved with the project.}

\item{abstract}{(character) Project abstract. Can pass as a character vector for separate paragraphs.}

\item{funding}{(character) Funding sources for the project such as grant and contract numbers. Can pass as a character vector for separate paragraphs.}

\item{studyAreaDescription}{(studyAreaDescription)}

\item{designDescription}{(designDescription)}

\item{relatedProject}{(project)}
}
\value{
(project) The new project section.
}
\description{
Note - studyAreaDescription, designDescription, and relatedProject are not fully fleshed out. Need to pass these objects in directly if you want to use them.
}
\examples{
eml_project("Some title", list(personnel1, personnel2), c("Abstract paragraph 1", "Abstract paragraph 2"), "#1 Best Scientist Award")
}
