% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml.R
\name{eml_validate_attributes}
\alias{eml_validate_attributes}
\title{Validate an EML attributeList attribute-by-attribute}
\usage{
eml_validate_attributes(attributes)
}
\arguments{
\item{attributes}{(attributeList) An attributeList.}
}
\value{
(logical) Named vector indicating which attributes are valid.
}
\description{
The attributes passed into this function are validated one-by-one and the
progress of going through each attribute is printed to the screen along
with any and all validation issues. This is done by, for each attribute in the list,
creating a minimum valid EML document and adding a new otherEntity with a new
attributeList containing the single attribute to be validated.
}
\examples{
\dontrun{
atts_df <- read.csv('attributes_table.csv', stringsAsFactors = F)
enum_domain <- read.csv('enumerated_domain.csv') # optional
attributes <- EML::set_attributes(atts_df, factor = enum_domain)
eml_validate_attributes(attributes)
}
}
