% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access.R
\name{filter_obsolete_pids}
\alias{filter_obsolete_pids}
\title{Filters PIDs that are obsolete.}
\usage{
filter_obsolete_pids(mn, pids)
}
\arguments{
\item{mn}{(MNode) The Member Node to query.}

\item{pids}{(character) PIDs to check the obsoletion state of.}
}
\value{
(character) PIDs that are not obsoleted by another PID.
}
\description{
Whether or not a PID is obsolete is determined by whether its "obsoletedBy"
property is set to another PID (TRUE) or is NA (FALSE).
}

