% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packaging.R
\name{filter_packaging_statements}
\alias{filter_packaging_statements}
\title{Filter statements related to packaging}
\usage{
filter_packaging_statements(statements)
}
\arguments{
\item{statements}{(data.frame) A set of Statements to be filtered}
}
\value{
(data.frame) The filtered Statements
}
\description{
This function was written specifically for the case of updating a resource
map while preserving any extra statements that have been added such as PROV
statements. Statements are filtered according to these rules:
}
\details{
1. If the subject or object is the ore:ResourceMap resource
2. If the subject or object is the ore:Aggregation resource
3. If the predicate is cito:documents or cito:isDocumentedBy
4. Once filters 1-3 have been executed, any remaining triples are considered
   for removal if they look like dangling dc:identifier statements

The consequence of filter 4 is that dc:identifier statements are left in if
they are still in use by another statement
}

