% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataone_formats.R
\name{format_eml}
\alias{format_eml}
\title{Helper function to generate the EML 2.1.1 format ID.}
\usage{
format_eml()
}
\value{
(character) The format ID for EML 2.1.1.
}
\description{
Helper function to generate the EML 2.1.1 format ID.
}
\examples{
format_eml

\dontrun{
# Upload a local EML 2.1.1 file:
env <- env_load()
publish_object(env$mn, "path_to_some_EML_file", format_eml())
}
}
