% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataone_formats.R
\name{format_iso}
\alias{format_iso}
\title{dataone_formats.R}
\usage{
format_iso()
}
\value{
(character) The format ID for ISO 19139.
}
\description{
A set of thin functions which return the DataONE format ID string. These are
to aid in filling in function arguments and can't remember or don't want to
type in the full format ID. By putting these format ID strings into
functions, a user's autocompletion routine in their editor can help them
fill in the format ID they want.
Helper function to generate the ISO 19139 format ID.w
}
\examples{
format_iso()
\dontrun{
# Upload a local ISO19139 XML file:
env <- env_load()
publish_object(env$mn, "path_to_some_EML_file", format_iso())
}
}

