% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packaging.R
\name{generate_resource_map}
\alias{generate_resource_map}
\title{Create a resource map RDF/XML file and save is to a temporary path.
This is a convenience wrapper around the constructor of the `ResourceMap`
class from `DataPackage`.}
\usage{
generate_resource_map(metadata_pid, data_pids = NULL, child_pids = NULL,
  resolve_base = "https://cn.dataone.org/cn/v2/resolve",
  resource_map_pid = NULL)
}
\arguments{
\item{metadata_pid}{(character) PID of the metadata Object.}

\item{data_pids}{(character) PID(s) of the data Objects.}

\item{child_pids}{(character) Optional. PID(s) of child Resource Maps.}

\item{resolve_base}{(character) Optional. The resolve service base URL.}

\item{resource_ma_pid}{(character) Optional. The PID to use for the resource map.}
}
\value{
Absolute path to the Resource Map on disk (character)
}
\description{
Create a resource map RDF/XML file and save is to a temporary path.
This is a convenience wrapper around the constructor of the `ResourceMap`
class from `DataPackage`.
}

