% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_all_sysmeta}
\alias{get_all_sysmeta}
\title{Get system metadata for all objects in a data package}
\usage{
get_all_sysmeta(node, rm_pid, nmax = 1000, child_packages = FALSE)
}
\arguments{
\item{node}{(MNode/CNode) The Coordinating/Member Node to run the query on.}

\item{rm_pid}{(character) The resource map PID of the package.}

\item{nmax}{(numeric) Default 1000. The maximum number of object system metadata to return. The default is sufficient to get system metadata for all objects in most packages.}

\item{child_packages}{(logical) Default FALSE. If parent package, whether or not to include objects of child packages.}
}
\value{
(list) A structured list of system metadata.
}
\description{
This function retrieves the system metadata for all objects in a data package and returns them as a list.
It is useful for inspecting system metadata for an entire data package and identifying changes where needed.
}
\examples{
\dontrun{
# Set environment
cn_staging <- CNode("STAGING2")
knb_test <- getMNode(cn_staging, "urn:node:mnTestKNB")
rm_pid <- "resource_map_urn:uuid:3e5307c4-0bf3-4fd3-939c-112d4d11e8a1"

all <- get_all_sysmeta(knb_test, rm_pid)

# View in RStudio data viewer to inspect
View(all)

# Print specific elements to console
all[[1]]@rightsHolder

# Create separate object
sysmeta_md <- all[[2]]
}
}
\keyword{internal}
