% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sysmeta.R
\name{get_all_sysmeta}
\alias{get_all_sysmeta}
\title{Get system metadata for all elements of a data package}
\usage{
get_all_sysmeta(mn, resource_map_pid, nmax = 1000, child_packages = FALSE)
}
\arguments{
\item{mn}{(MNode) The Member Node to query.}

\item{resource_map_pid}{(character) The PID for a resource map.}

\item{nmax}{(numeric) The maximum number of system metadata objects to return.}

\item{child_packages}{(logical) If parent package, whether or not to include child packages.}
}
\value{
(list) A list of system metadata objects.
}
\description{
This function retrieves the system metadata for all elements of a data package and returns them as a list.
It is useful for inspecting system metadata for an entire data package and identifying changes where needed.
}
\examples{
\dontrun{
cn_staging <- CNode("STAGING")
adc_test <- getMNode(cn_staging, "urn:node:mnTestARCTIC")

rm_pid <- "resource_map_urn:uuid:..."

all <- get_all_sysmeta(adc_test, rm_pid)

# View in viewer to inspect
View(all)

# Print specific elements to console
all[[1]]@rightsHolder

# Create separate object
sysmeta_md <- all[[2]]
}
}
