% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataone.R
\name{get_mn_base_url}
\alias{get_mn_base_url}
\title{Get base URL of a Member Node}
\usage{
get_mn_base_url(mn)
}
\arguments{
\item{mn}{(character) The Member Node.}
}
\value{
(character) The URL.
}
\description{
Get the base URL of a Member Node.
}
\examples{
\dontrun{
cn <- CNode('STAGING2')
mn <- getMNode(cn,"urn:node:mnTestKNB")
}
}
