% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get_package}
\alias{get_package}
\title{Get a structured list of PIDs for the objects in a package}
\usage{
get_package(node, pid, file_names = FALSE, rows = 5000)
}
\arguments{
\item{node}{(MNode/CNode) The Coordinating/Member Node to run the query on.}

\item{pid}{(character) The the resource map PID of the package.}

\item{file_names}{(logical) Whether to return file names for all objects.}

\item{rows}{(numeric) The number of rows to return in the query. This is only
useful to set if you are warned about the result set being truncated. Defaults to 5000.}
}
\value{
(list) A structured list of the members of the package.
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("superseded")}
}
\details{
Please use dataone::getDataPackage() when possible

Get a structured list of PIDs for the objects in a package,
including the resource map, metadata, and data objects.
}
\examples{
\dontrun{
#Set environment
cn <- CNode("STAGING2")
mn <- getMNode(cn,"urn:node:mnTestKNB")
pid <- "resource_map_urn:uuid:3e5307c4-0bf3-4fd3-939c-112d4d11e8a1"

ids <- get_package(mn, pid)
}
}
