% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access.R
\name{get_package}
\alias{get_package}
\title{access.R}
\usage{
get_package(mn, pid)
}
\arguments{
\item{mn}{(MNode) The Member Node to run the query on.}

\item{pid}{(character) The the metadata PID of the package.}
}
\description{
Utility functions for getting and setting access rules for DataONE objects.
Get a structured list of PIDs for the objects in a package.
}

