% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access.R
\name{get_related_pids}
\alias{get_related_pids}
\title{Get the related PIDs for a given PID (i.e. data files, resource maps).}
\usage{
get_related_pids(mn, pid, flatten = TRUE)
}
\arguments{
\item{flatten}{(logical) Whether to flatten the result into a single character vector.}
}
\value{
Either a character vector or list
}
\description{
Get the related PIDs for a given PID (i.e. data files, resource maps).
}

