% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{guess_format_id}
\alias{guess_format_id}
\title{Guess format from filename for a vector of filenames.}
\usage{
guess_format_id(filenames)
}
\arguments{
\item{filenames}{(character)}
}
\value{
(character) DataOne format identifiers strings.
}
\description{
Guess format from filename for a vector of filenames.
}
