% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inventory.R
\name{inv_load_files}
\alias{inv_load_files}
\title{Load files into the inventory from a text file.}
\usage{
inv_load_files(inventory, path)
}
\arguments{
\item{inventory}{(character) A \code{data.frame}.}

\item{path}{(character) Path to a file containing a file listing.}
}
\value{
An inventory (data.frame)
}
\description{
Files should be the output of the command:
}
\details{
you@server:/path/to/acadis$ find . -type f
}

