% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataone.R
\name{is_authorized}
\alias{is_authorized}
\title{Check if the user has authorization to perform an action on an object.}
\usage{
is_authorized(node, ids, action)
}
\arguments{
\item{node}{(MNode|CNode) The Node to query.}

\item{ids}{(character) The PID or SID to check.}

\item{action}{(character) One of read, write, or changePermission.}
}
\value{
(boolean)
}
\description{
Check if the user has authorization to perform an action on an object.
}
\examples{
\dontrun{
cn <- CNode('STAGING2')
mn <- getMNode(cn,"urn:node:mnTestKNB")
pids <- c("urn:uuid:3e5307c4-0bf3-4fd3-939c-112d4d11e8a1", "urn:uuid:23c7cae4-0fc8-4241-96bb-aa8ed94d71fe")
is_authorized(mn, pids, "write")
}
}
