% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{is_obsolete}
\alias{is_obsolete}
\title{Test whether the object is obsoleted by another object}
\usage{
is_obsolete(node, pids)
}
\arguments{
\item{node}{(MNode|CNode) The Coordinating/Member Node to run the query on.}

\item{pids}{(character) One or more PIDs to query against.}
}
\value{
(logical) Whether or not the object is obsoleted by another object.
}
\description{
Test whether the object is obsoleted by another object
}
\examples{
\dontrun{
# Set environment
cn <- CNode("STAGING2")
mn <- getMNode(cn,"urn:node:mnTestKNB")
pid <- "urn:uuid:3e5307c4-0bf3-4fd3-939c-112d4d11e8a1"

is_obsolete(mn, pid)
}
}
