% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{is_resource_map}
\alias{is_resource_map}
\title{Determines whether the object with the given PID is a resource map.}
\usage{
is_resource_map(node, pids)
}
\arguments{
\item{node}{(MNode|CNode) The Coordinating/Member Node to run the query on.}

\item{pids}{(character) Vector of PIDs}
}
\value{
(logical) Whether or not the object(s) are resource maps
}
\description{
Determines whether the object with the given PID is a resource map.
}

