% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{is_resource_map}
\alias{is_resource_map}
\title{Determines whether the object with the given PID is a resource map.}
\usage{
is_resource_map(mn, pid)
}
\arguments{
\item{mn}{(MNode)}

\item{pid}{(character)}
}
\value{
(logical)
}
\description{
Determines whether the object with the given PID is a resource map.
}

