% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{object_exists}
\alias{object_exists}
\title{Check if an object exists on a Member Node}
\usage{
object_exists(node, pids)
}
\arguments{
\item{node}{(MNode) The Member Node to query.}

\item{pids}{(character) The PID(s) to check the existence of.}
}
\value{
(logical) Whether the object exists.
}
\description{
This is a simple check for the HTTP status of a /meta/\{PID\} call on the
provided Member Mode.
}
\examples{
\dontrun{
# Set environment
cn <- CNode("STAGING2")
mn <- getMNode(cn,"urn:node:mnTestKNB")
pids <- c("urn:uuid:3e5307c4-0bf3-4fd3-939c-112d4d11e8a1",
"urn:uuid:23c7cae4-0fc8-4241-96bb-aa8ed94d71fe")

object_exists(mn, pids)
}
}
