% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml.R
\name{pid_to_eml_entity}
\alias{pid_to_eml_entity}
\title{Create EML entity with physical section from any DataONE PID}
\usage{
pid_to_eml_entity(mn, pid, entity_type = "otherEntity", ...)
}
\arguments{
\item{mn}{(MNode) Member Node where the PID is associated with an object.}

\item{pid}{(character) The PID of the object to create the sub-tree for.}

\item{entity_type}{(character) What kind of object to create from the input. One of "dataTable",
"spatialRaster", "spatialVector", "storedProcedure", "view", or "otherEntity".}

\item{...}{(optional) Additional arguments to be passed to \code{eml$entityType())}.}
}
\value{
(list) The entity object.
}
\description{
Create EML entity with physical section from any DataONE PID
}
\examples{
\dontrun{
# Generate EML otherEntity
pid_to_eml_entity(mn,
                  pid,
                  entity_type = "otherEntity",
                  entityName = "Entity Name",
                  entityDescription = "Description about entity")
}
}
