% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml.R
\name{pid_to_eml_other_entity}
\alias{pid_to_eml_other_entity}
\title{eml.R}
\usage{
pid_to_eml_other_entity(mn, pids)
}
\arguments{
\item{mn}{(MNode) Member Node where the PID is associated with an object.}

\item{pids}{(character) The PID of the object to create the sub-tree for.}
}
\value{
(list of otherEntity) The otherEntity object(s)
}
\description{
Helpers for creating EML.
Create EML otherEntity objects for a set of PIDs
}
\details{
Note this is a wrapper around sysmeta_to_other_entity which handles the task of
creating the EML otherEntity.
}
\examples{
\dontrun{
Generate EML otherEntity objects for all the data in a package
pkg <- get_package(mn, pid)
pid_to_other_entity(mn, pkg$data)
}
}
