% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml.R
\name{pid_to_eml_physical}
\alias{pid_to_eml_physical}
\title{Create EML physical objects for the given set of PIDs}
\usage{
pid_to_eml_physical(mn, pids)
}
\arguments{
\item{mn}{(MNode) Member Node where the PID is associated with an object.}

\item{pids}{(character) The PID of the object to create the sub-tree for.}
}
\value{
(list) A list of otherEntity object(s).
}
\description{
This is a wrapper around \code{\link[=sysmeta_to_eml_physical]{sysmeta_to_eml_physical()}} which handles the task of
creating the EML physical.
}
\examples{
\dontrun{
# Generate EML physical objects for all the data in a package
pkg <- get_package(mn, pid)
pid_to_eml_physical(mn, pkg$data)
}
}
