% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml.R
\name{pid_to_eml_physical}
\alias{pid_to_eml_physical}
\title{Create EML physical objects for the given set of PIDs}
\usage{
pid_to_eml_physical(mn, pid, headers = 1)
}
\arguments{
\item{mn}{(MNode) Member Node where the PID is associated with an object.}

\item{pid}{(character) The PID of the object to create the physical for.}

\item{headers}{(double) The number of headers in a csv file. Default is equal to 1.}
}
\value{
(list) A physical object.
}
\description{
This function creates a data object's physical.
}
\examples{
\dontrun{
# Generate EML physical sections for an object in a data package
phys <- pid_to_eml_physical(mn, pid, headers)
}
}
