% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editing.R
\name{publish_update}
\alias{publish_update}
\title{Publish an updated data package.}
\usage{
publish_update(mn, metadata_pid, resource_map_pid, data_pids = NULL,
  child_pids = NULL, metadata_path = NULL, identifier = NULL,
  use_doi = FALSE, parent_resmap_pid = NULL, parent_metadata_pid = NULL,
  parent_data_pids = NULL, parent_child_pids = NULL, public = TRUE,
  check_first = TRUE)
}
\arguments{
\item{mn}{(MNode) The Member Node to update the object on.}

\item{metadata_pid}{(character) The PID of the EML metadata document to be updated.}

\item{resource_map_pid}{(character)  The PID of the resource map for the package.}

\item{data_pids}{(character)  PID(s) of data objects that will go in the updated package.}

\item{child_pids}{(character) Optional. Child packages resource map PIDs.}

\item{metadata_path}{(character) Optional. Path to a metadata file to update with. If this is not set, the existing metadata document will be used.}

\item{identifier}{(character) Manually specify the identifier for the new metadata object.}

\item{use_doi}{(logical) Generate and use a DOI as the identifier for the updated metadata object.}

\item{parent_resmap_pid}{(character)  Optional. PID of a parent package to be updated.}

\item{parent_metadata_pid}{(character)  Optional. identifier for the metadata document of the parent package.}

\item{parent_child_pids}{(character) Optional. Resource map identifier(s) of child packages in the parent package.}

\item{public}{(logical) Optional. Make the update public.
This applies to the new metadata PID and its resource map and data object.
access policies are not affected.}

\item{check_first}{(logical) Optional. Whether to check the PIDs passed in as aruments exist on the MN before continuing. This speeds up the function, especially when `data_pids` has many elements.}

\item{parent_data_pids(character)}{Optional. Identifier(s) of data in the parent package.}
}
\description{
This function can be used for a variety of tasks:
}
\details{
\itemize{
  \item Publish an existing package with a DOI
  \item Update a package with new data objects
  \item Update a package with new metadata
}

The metadata_pid and resource_map_pid provide the identifier of an EML metadata
document and associated resource map, and the data_pids vector provides a list
of PIDs of data objects in the package.  Update the metadata file and resource map
by generating a new identifier (a DOI if use_doi is TRUE) and updating the Member
Node with a public version of the object.  If metadata_file is not missing, it
should be an edited version of the metadata to be used to update the original. If
parent_resmap_pid is not missing, it indicates the PID of a parent package that
should be updated as well, using the parent_medata_pid, parent_data_pids, and
parent_child_pids as members of the updated package. In all cases, the objects
are made publicly readable.
}

