% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access.R
\name{remove_public_read}
\alias{remove_public_read}
\title{Remove public access on a set of objects.}
\usage{
remove_public_read(mn, pids)
}
\arguments{
\item{mn}{(MNode)}

\item{pids}{(character) A vector of PIDs to set public access on}
}
\description{
Remove public access on a set of objects.
}

