% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editing.R
\name{set_file_name}
\alias{set_file_name}
\title{Set the file name on an object}
\usage{
set_file_name(mn, pid, name)
}
\arguments{
\item{mn}{(MNode) The Member Node.}

\item{pid}{(character) The PID of the object to set the file name on.}

\item{name}{(character) The file name.}
}
\value{
(logical) Whether the update succeeded, FALSE means there was an error.
}
\description{
Set the file name on an object
}

