% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml.R
\name{set_other_entities}
\alias{set_other_entities}
\title{Creates and sets EML otherEntity elements to an existing EML document.}
\usage{
set_other_entities(mn, path, pids)
}
\arguments{
\item{mn}{(MNode) The Member Node the objects exist on.}

\item{path}{(character) The location on disk of the EML file.}

\item{pids}{(character) One or more PIDs for the objects.}
}
\value{
(character) The path to the updated EML file.
}
\description{
This function isn't that smart. It will remove existing otherEntity elements
if what's in their 'id' attribute isn't `pids`. It will then go on to add
any otherEntity elements it needs. If your EML document doesn't store the
PID for the otherEntity in the 'id' attribute, you may be confused when your
otherEntity is removed and then added back. In theory, this isn't too bad
but, in practice, this slows down the execution of the function because each
new otherEntity element requires a network call to find its 'docid'.
}

