% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access.R
\name{set_public_read}
\alias{set_public_read}
\title{Set public access on a set of objects.}
\usage{
set_public_read(mn, pids)
}
\arguments{
\item{mn}{(MNode)}

\item{pids}{(character) A vector of PIDs to set public access on}
}
\description{
Set public access on a set of objects.
}
