% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access.R
\name{set_rights_holder}
\alias{set_rights_holder}
\title{Set the rights holder for an object}
\usage{
set_rights_holder(mn, pids, subject)
}
\arguments{
\item{mn}{(MNode) The Member Node.}

\item{pids}{(character) The PIDs of the objects to set the rights holder for.}

\item{subject}{(character) The identifier of the new rights holder, typically an ORCID or DN.}
}
\value{
(logical) Whether an update was needed.
}
\description{
Set the rights holder to the given subject for the given objects on the
given Member Node. This function checks if the rights holder is already set
and only updates the System Metadata when a change is needed.
}
\examples{
\dontrun{
cn <- CNode("STAGING2")
mn <- getMNode(cn,"urn:node:mnTestKNB")
pids <- c("urn:uuid:3e5307c4-0bf3-4fd3-939c-112d4d11e8a1",
"urn:uuid:23c7cae4-0fc8-4241-96bb-aa8ed94d71fe")
set_rights_holder(mn, pids, subjects = "http://orcid.org/0000-000X-XXXX-XXXX")
}
}
