% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml.R
\name{sysmeta_to_eml_other_entity}
\alias{sysmeta_to_eml_other_entity}
\title{This function is deprecated. See \link{pid_to_eml_entity}.}
\usage{
sysmeta_to_eml_other_entity(sysmeta)
}
\arguments{
\item{sysmeta}{(SystemMetadata) One or more System Metadata objects}
}
\description{
This function is deprecated. See \link{pid_to_eml_entity}.
}
