% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml.R
\name{sysmeta_to_eml_physical}
\alias{sysmeta_to_eml_physical}
\title{Create an EML physical subtree from a System Metadata instance}
\usage{
sysmeta_to_eml_physical(sysmeta)
}
\arguments{
\item{sysmeta}{}
}
\description{
This function creates a pre-canned EML physical subtree from what's in the
System Metadata of an Object. Note that it sets an Online Distrubtion URL
of the DataONE v2 resolve service for the PID.
}

