% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml.R
\name{sysmeta_to_other_entity}
\alias{sysmeta_to_other_entity}
\title{Create an EML otherEntity sub-tree for the given object.}
\usage{
sysmeta_to_other_entity(sysmeta)
}
\arguments{
\item{sysmeta}{(SystemMetadata) The System Metadata of the object.}
}
\value{
(otherEntity) The XML sub-tree.
}
\description{
Create an EML otherEntity sub-tree for the given object.
}
