% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editing.R
\name{update_resource_map}
\alias{update_resource_map}
\title{Update an existing Resource Map with a new one.}
\usage{
update_resource_map(mn, old_resource_map_pid, new_resource_map_pid = NULL,
  metadata_pid, data_pids = NULL, child_pids = NULL, public = FALSE,
  check_first = TRUE)
}
\arguments{
\item{public}{Whether or not to make the new resource map public read (logical)}

\item{check_first}{(logical) Optional. Whether to check the PIDs passed in as aruments exist on the MN before continuing. This speeds up the function, especially when `data_pids` has many elements.}
}
\description{
This function is intended to be used to add a few new child packages to a
parent package. For exmaple, if you have:
}
\details{
Parent
  A
  B

and want to add C as a sibling package to A and B, e.g.

Parent
  A
  B
  C

you could use this function.

Note: This function currently replaces the rightsHolder on the Resource Map
temporarily to allow updating but sets it back to the rightsHolder that was
in place before the update.
}

